Reset to master state, as the idea for the integration of cuda code changed -> use extra library
