+++ .gitignore
@@ -74,7 +74,3 @@ doc/html/
 /tests/test_enum
 /tests/test_bkz_gram
 /tests/test_svp_gram
-
-.vscode
-fplll/mybench
-fplll/mybench_dbg
