+++ .travis-check-style.sh
