+++ doc/templates/jquery.smartmenus.js
