--- fplll/enum-cuda/CudaEnumeration.vcxproj.filters
@@ -1,46 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="unittest">
-      <UniqueIdentifier>{913d1846-5221-494e-a769-5662bf05a617}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="utility">
-      <UniqueIdentifier>{00695613-389d-4852-90e3-54fc6a1b93d8}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="testdata.h">
-      <Filter>unittest</Filter>
-    </ClInclude>
-    <ClInclude Include="util.h">
-      <Filter>utility</Filter>
-    </ClInclude>
-    <ClInclude Include="prefix.cuh">
-      <Filter>utility</Filter>
-    </ClInclude>
-    <ClInclude Include="recenum.cuh" />
-    <ClInclude Include="types.cuh" />
-    <ClInclude Include="cuda_util.cuh">
-      <Filter>utility</Filter>
-    </ClInclude>
-    <ClInclude Include="atomic.h">
-      <Filter>utility</Filter>
-    </ClInclude>
-    <ClInclude Include="enum.cuh" />
-    <ClInclude Include="wrapper.h" />
-    <ClInclude Include="cuda_check.h">
-      <Filter>utility</Filter>
-    </ClInclude>
-  </ItemGroup>
-  <ItemGroup>
-    <CudaCompile Include="wrapper.cu" />
-    <CudaCompile Include="test.cpp">
-      <Filter>unittest</Filter>
-    </CudaCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="testdata.cpp">
-      <Filter>unittest</Filter>
-    </ClCompile>
-  </ItemGroup>
-</Project>
