--- fplll/enum-cuda/testdata.cpp
@@ -1,286 +1,0 @@
-#include "testdata.h"
-
-const std::array<std::array<float, 60>, 60> test_mu_big = { { 
- {-7.93925659e+02, -3.69873413e+02, 3.84824432e+02, -3.42583710e+02, -2.69462250e+02, -2.39153366e+02, 3.74196228e+02, 3.70014465e+02, -1.34457169e+02, -3.30501953e+02, 4.31614685e+01, 3.19419067e+02, -3.69175598e+02, -3.77046631e+02, 3.82261230e+02, -3.54792664e+02, 2.71837799e+02, -6.54696045e+01, 3.92593414e+02, -3.26995331e+02, 7.39439468e+01, -1.30226297e+01, 3.65071960e+02, 9.31044312e+01, 3.68911102e+02, -3.92662689e+02, 2.98817627e+02, 3.31726257e+02, 3.52723175e+02, 8.81568623e+00, -3.84271484e+02, 3.65976318e+02, 3.12614899e+02, -2.74438782e+02, -3.26922272e+02, -2.03027817e+02, -3.33577820e+02, 3.84577545e+02, -3.96384705e+02, -9.29003830e+01, 3.30271454e+02, 3.74895294e+02, 2.42273300e+02, -3.92056854e+02, -2.46487808e+02, -3.38660156e+02, 4.23112640e+01, -2.01007477e+02, 1.70714966e+02, 2.98645081e+02, 2.49471710e+02, 5.43892212e+01, 1.93610062e+02, -3.64292267e+02, -3.24026550e+02, 7.59139061e+00, -1.72159683e+02, 7.87428894e+01, 1.84483261e+02, -2.04030426e+02},
- { 0.00000000e+00, 6.48447144e+02, 2.35441437e+02, 2.71712830e+02, 1.76282806e+02, 1.91472092e+02, 2.26024963e+02, 2.39281677e+02, -3.07565765e+02, 1.29791794e+02, -3.17010773e+02, 2.29254807e+02, 1.85428787e+02, 1.34733398e+02, 1.11142845e+02, 2.32683014e+02, 2.66241547e+02, -3.19891113e+02, 1.91168808e+02, 2.34546677e+02, -3.00687744e+02, -2.21362274e+02, 2.61218536e+02, -2.51710800e+02, 1.38027298e+02, 2.57721100e+02, 1.12241531e+02, 2.79951477e+02, 2.50186829e+02, -2.33522995e+02, 2.72397217e+02, 2.81973511e+02, 3.05901489e+02, 2.46062302e+02, 1.90897827e+02, 2.41738159e+02, 3.22420166e+02, 1.89372437e+02, 2.03834412e+02, -3.01838623e+02, 1.82134567e+02, 2.43061920e+02, 1.18690519e+01, 2.14707718e+02, 4.91138229e+01, 2.08842957e+02, -2.96838684e+02, 5.96527901e+01, -1.25683456e+02, 2.85481842e+02, 2.82127777e+02, -2.51335648e+02, -2.66275818e+02, 8.48056412e+01, -4.78879509e+01, 3.19390503e+02, -2.95275116e+02, 2.94108643e+02, 2.24896454e+02, 1.13276115e+02},
- { 0.00000000e+00, 0.00000000e+00, 6.40563416e+02, 2.35919052e+02, 1.66524307e+02, 1.98644806e+02, 2.80716156e+02, 1.97661255e+02, 1.36572159e+02, 1.06150093e+02, 1.48275543e+02, 5.60032921e+01, -2.22566986e+01, 3.39637375e+01, 3.05051636e+02, 2.09319870e+02, 2.43830475e+02, 1.01426216e+02, 1.15366257e+02, 1.52548187e+02, 1.68154968e+02, 5.44681091e+01, 2.19706161e+02, 1.93509796e+02, 1.63172989e+02, 1.46776533e+01, 3.03356964e+02, 2.46151199e+02, 3.65071640e+01, 1.54672424e+02, 1.00660538e+02, 1.10749573e+02, 3.86897888e+01, 7.21684723e+01, 2.12374634e+02, 1.22607506e+02, 1.53737823e+02, 2.22304367e+02, 1.52772461e+02, 2.23377808e+02, 2.65930664e+02, 2.22421829e+02, -1.27275696e+02, 2.81271057e+02, -2.13332352e+02, -2.01008081e+00, 2.18099747e+02, -2.35744263e+02, 1.83085388e+02, -2.68902618e+02, -2.30911377e+02, 1.99475220e+02, -1.72239807e+02, 2.12673416e+02, 2.58290424e+01, -1.96550583e+02, -2.97813995e+02, -3.06028320e+02, 2.94149475e+02, 2.80932373e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 6.09172180e+02, 1.39643723e+02, 9.66137924e+01, 2.17597519e+02, 2.06250336e+02, 1.10160568e+02, 1.55247559e+02, 1.43375366e+02, 1.75720520e+02, 2.04444427e+02, 2.02315350e+01, 1.39522186e+02, 2.32003891e+02, 1.64524399e+02, 1.23494347e+02, 1.10817581e+02, 1.02231133e+02, 1.12783791e+02, 2.52441864e+02, 1.82932526e+02, 1.46657166e+02, 1.10956352e+02, 1.77523361e+02, 1.00802208e+02, 2.04726761e+02, 1.91792252e+02, 1.49480118e+02, 2.99301727e+02, 2.48720764e+02, 1.44346298e+02, 1.66530167e+02, 1.55604843e+02, 1.90916748e+02, 1.93554962e+02, 1.72848083e+02, 1.94097336e+02, 3.18823504e+00, -2.51452503e+01, 1.71548264e+02, 2.73252258e+02, 1.61438217e+02, -1.74673752e+02, 1.30229935e+02, 1.04299149e+02, 2.01183273e+02, -2.77234192e+02, -2.79393749e+01, -1.47940765e+02, -2.95021729e+02, 1.45218155e+02, -1.56937714e+02, -2.93870941e+02, 2.73202271e+02, 2.02742233e+02, 8.13889008e+01, 1.34694656e+02, -1.47843872e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.71671875e+02, -7.73752594e+01, -3.15744934e+01, -1.19826981e+02, -1.48063278e+02, -3.66717720e+01, -5.95970993e+01, -1.92347260e+02, -1.15188637e+02, -1.18616982e+02, -6.44941406e+01, -9.00009766e+01, -1.73581573e+02, -1.21040718e+02, -1.08721565e+02, -2.30185593e+02, -1.47505493e+02, 3.29838295e+01, -9.11606445e+01, -4.67778473e+01, -1.72501297e+02, -2.37465210e+01, -5.56282692e+01, -8.54670486e+01, -1.10281845e+02, -1.94113216e+01, 8.02495193e+00, -1.12650162e+02, -3.89067497e+01, -1.53426926e+02, -6.16787453e+01, -8.11408310e+01, -2.58809624e+01, -2.54045715e+02, -1.53420135e+02, -1.34365112e+02, -4.19558105e+01, -7.73279419e+01, 8.70696411e+01, -1.38663284e+02, 2.04522324e+01, -7.99083939e+01, -4.57642984e+00, -2.57029388e+02, -2.10538921e+01, -1.46444855e+02, 1.63427170e+02, -2.62168827e+01, 2.17530792e+02, -2.41887497e+02, -1.57775757e+02, 1.54700623e+02, 1.83432175e+02, 1.14482376e+02, -1.61389069e+02, -5.17189484e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.79056824e+02, -1.02050369e+02, -6.55878878e+00, -4.16173744e+01, -1.45225143e+02, -1.90321899e+02, -7.10770798e+00, -1.50653442e+02, -3.39352455e+01, -2.96828823e+01, 4.18243217e+01, -1.44999924e+02, -1.93819321e+02, -1.42891968e+02, -7.87537994e+01, -2.05357483e+02, -1.15771240e+02, -1.18070023e+02, -1.50711334e+02, 3.21471748e+01, -6.12213173e+01, 2.21075668e+01, -2.11792923e+02, -2.32967712e+02, -1.34082687e+02, 7.77066469e+00, -2.21531372e+02, -1.64620270e+02, -6.63810806e+01, -1.03183891e+02, -1.79142990e+02, -9.58468094e+01, -6.60894775e+01, -6.85329971e+01, 1.07674325e+00, -9.31823883e+01, -1.04504883e+02, 2.85430641e+01, -1.13356415e+02, -9.34979782e+01, -1.21989838e+02, 2.40591064e+02, -8.76334839e+01, -2.08229355e+02, -5.88071480e+01, 2.05848633e+02, -1.43362137e+02, -3.33852539e+02, 1.96367752e+02, -3.06016235e+02, 3.22934998e+02, -3.11823669e+02, -2.31291061e+02, 1.57263336e+02, -3.10484253e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.34145447e+02, -8.11206207e+01, -1.54056351e+02, -4.22737465e+01, 1.27812071e+01, -1.91436050e+02, -4.08952599e+01, -1.56564045e+01, 6.83894825e+00, -7.39947205e+01, -1.30433456e+02, -8.43618469e+01, -1.53297472e+01, -4.57074776e+01, -9.21663666e+01, 2.22227173e+01, -2.39982319e+01, -5.80135012e+00, -3.40806808e+01, -1.92047684e+02, -1.57585800e+02, -3.39787712e+01, -1.16264565e+02, -3.71469765e+01, -1.07214195e+02, -7.38120041e+01, -1.24805824e+02, -1.16241058e+02, -3.72213707e+01, -1.25754875e+02, -3.70426140e+01, -1.91676956e+02, -1.19638512e+02, -4.43492737e+01, -9.87773132e+01, -6.31516113e+01, 1.80310562e+02, -1.90690826e+02, -8.79536343e+00, -1.87986862e+02, 9.91447906e+01, -1.87377945e+02, -2.07629807e+02, 2.50283676e+02, 2.59268045e+00, 3.14856781e+02, 2.95935486e+02, 2.56986908e+02, -2.50963867e+02, -3.58105965e+01, -8.14530945e+01, -2.46598450e+02, 3.12431305e+02, 2.65733429e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.20485535e+02, -1.89791992e+02, -2.30796753e+02, 3.63503189e+01, -4.81241341e+01, -1.79486942e+01, -8.73610306e+01, -1.45417433e+01, -8.32390060e+01, -5.08338699e+01, -9.86864090e+00, -1.26790092e+02, -1.75637329e+02, -4.30270996e+01, 1.55639477e+01, -3.68555565e+01, -1.28958649e+02, -6.80084686e+01, -8.24135971e+01, -1.89999142e+01, -7.00789261e+01, -1.24156609e+01, -6.01270523e+01, -1.23849556e+02, -1.00688156e+02, -2.13724213e+01, -7.23138428e+01, -1.37986746e+01, -5.06348419e+01, -3.47418213e+01, -8.34902496e+01, -4.79155121e+01, -3.94079781e+01, -1.56902710e+02, 2.51968956e+01, 4.10197639e+01, -7.97971344e+01, 1.09133743e+02, -1.00823166e+02, -1.44456177e+02, 3.48949242e+01, 2.55173035e+01, -1.07947283e+01, -2.59574677e+02, 2.91293964e+01, -1.36066193e+02, 7.34370499e+01, -9.68073654e+01, 1.56650635e+02, -2.31598541e+02, 2.45279678e+02, 1.12048599e+02, 3.07629333e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.17478882e+02, 1.15512779e+02, -1.30678604e+02, -5.39844627e+01, -9.33626556e+01, -8.35510864e+01, -5.63893509e+01, -7.22782516e+01, -4.40538073e+00, -3.72789764e+01, -2.03502440e+00, -1.67526932e+02, 5.65074768e+01, -1.03194878e+02, -1.99194580e+02, -5.72160454e+01, 3.67068825e+01, -6.56796188e+01, -8.20719757e+01, -7.27917709e+01, -9.59763565e+01, -4.66627426e+01, -2.01873169e+01, -8.27624817e+01, -6.83015289e+01, 4.23965034e+01, -1.29805161e+02, 2.69876404e+01, 4.16390495e+01, -4.54378471e+01, 1.77459049e+01, -1.21965164e+02, -1.70604599e+02, 3.31913986e+01, 2.65377445e+01, 3.67892075e+01, 9.47596588e+01, -3.10967369e+01, -1.68587234e+02, 2.74820374e+02, -9.97581387e+00, 2.80548279e+02, -9.37393646e+01, 2.07037354e+01, 6.66978226e+01, -6.96000671e+01, -1.37121857e+02, -2.34472626e+02, -1.13247040e+02, -2.16087341e+02, -7.88511581e+01, 7.14100266e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.43263611e+02, 2.55234356e+01, 4.77180624e+00, -2.03418946e+00, 1.12636879e+02, -4.89002304e+01, 7.10869014e-01, -1.64590103e+02, -1.71994743e+01, -5.84582138e+01, -1.83346420e+02, 6.54761124e+00, -5.88678818e+01, 5.15156136e+01, -5.99472237e+01, -3.94133997e+00, -3.63392487e+01, 3.41487808e+01, 3.82069206e+01, -8.64167175e+01, -9.10215683e+01, -1.59330845e+01, -1.12228889e+01, 2.47702522e+01, 8.94268990e+00, 9.58799286e+01, 6.24916306e+01, -2.21072479e+02, -1.45179768e+01, -8.36886597e+01, -1.08251930e+02, -1.17820595e+02, -9.97266006e+01, 2.25392780e+01, -2.14432163e+01, -4.85792885e+01, 1.14333658e+01, -4.06523094e+01, -2.06875916e+02, 7.88596058e+00, 1.17972931e+02, 4.41075211e+01, -2.69031982e+01, -5.89494667e+01, -1.69382111e+02, -2.51614868e+02, 2.26421936e+02, 1.95533035e+02, -2.44379616e+01, -2.51638336e+02, -2.47759140e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.16421204e+02, -1.17054352e+02, -6.20441360e+01, 5.42237663e+00, 2.13793964e+01, -3.96481895e+01, -1.08917297e+02, -1.62278900e+01, 1.39302740e+01, -1.84442444e+02, -1.89600769e+02, -2.10214462e+01, -1.09015427e+02, 3.32729645e+01, 2.30097427e+01, -1.06100250e+02, 5.97694054e+01, 8.21882343e+00, -1.28844513e+02, -2.27301769e+01, 2.18316135e+01, 6.08340168e+00, 2.32708912e+01, 1.56475611e+01, -3.13302345e+01, -5.20256157e+01, 7.78786392e+01, 2.53444176e+01, 2.91617966e+01, -1.55300095e+02, 2.93181362e+01, -8.65642700e+01, -2.44096130e+02, -5.74418068e+01, -8.59449463e+01, -9.58971634e+01, 4.56804924e+01, 1.03032646e+02, -1.57034790e+02, -1.74641205e+02, -8.78755951e+01, -6.21035843e+01, 8.46316757e+01, 1.43891022e+02, 1.67724091e+02, -1.55529709e+02, -1.41961319e+02, -2.17922791e+02, -2.23709961e+02, 1.19102142e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.02703827e+02, -1.35636688e+02, 3.09387569e+01, -1.70489101e+01, -1.51478405e+01, 7.45803404e+00, 5.66459122e+01, 3.08483276e+01, -4.38579273e+00, -6.04819870e+01, 5.92605858e+01, -1.09757965e+02, -3.44782219e+01, 4.15189314e+00, -3.11959147e+00, -1.04190521e+02, -9.67214966e+01, -3.43978348e+01, -6.68117523e+01, -1.24409622e+02, -9.53506546e+01, -9.68265762e+01, -8.73703995e+01, -7.59240341e+01, 3.68894882e+01, -7.23120728e+01, -7.37746964e+01, -1.51174331e+00, -1.33528168e+02, -7.92128601e+01, -1.07830353e+02, 1.93306458e+02, 1.18112038e+02, -7.26473923e+01, 1.09496651e+01, -4.69256554e+01, -5.18209763e+01, -1.56811966e+02, -2.20931885e+02, 2.10606918e+02, -1.06769928e+02, -2.04987656e+02, 1.26781311e+02, -2.36790268e+02, -8.41104584e+01, 1.29245300e+02, -1.57114380e+02, 6.09450607e+01, -2.50291351e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.19520630e+02, -8.46379929e+01, -1.81294205e+02, -1.27572578e+02, -3.15554752e+01, -1.05398499e+02, 8.58837738e+01, 5.64595833e+01, -6.94305878e+01, 2.35952377e+01, 8.51017990e+01, 2.32623825e+01, -9.13983459e+01, -9.10464764e+00, -4.51805725e+01, 2.47872238e+01, -5.48399048e+01, 2.35965395e+00, -3.88883247e+01, -1.19903320e+02, 4.25867748e+00, -7.63999634e+01, -9.31207047e+01, -9.70426712e+01, -4.88411407e+01, -2.75767064e+00, -1.05273857e+02, -6.51268768e+01, -1.18873047e+02, -1.22437019e+02, 8.89902344e+01, -4.58117561e+01, 1.14775124e+02, 1.67529297e+02, -9.46524811e+01, -1.68453537e+02, -1.23541603e+02, -1.93067474e+01, -1.21783775e+02, 1.82245926e+02, -5.02274475e+01, -1.81569000e+02, 1.12792374e+02, 2.46385300e+02, 9.86144180e+01, 1.97927521e+02, 2.10363754e+02, -1.42876984e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.30021912e+02, 3.14737034e+01, 2.48558140e+01, -3.61125908e+01, -8.62666626e+01, -3.39697456e+01, -1.25296455e+02, 9.17353516e+01, 1.06635361e+01, 7.52193527e+01, 5.47613869e+01, -7.03996048e+01, -9.10284958e+01, -2.44395466e+01, -5.39516220e+01, -1.27227692e+02, -5.41628876e+01, -9.93321686e+01, -1.76945591e+01, -4.16687965e+00, 1.14002466e+01, -4.29378510e+01, -1.07280571e+02, 4.29210167e+01, -1.68824142e+02, 2.67136269e+01, -9.88735046e+01, 3.14653339e+01, -9.80732040e+01, -1.32775833e+02, 7.86185760e+01, -1.32900787e+02, -8.18680267e+01, -1.81115356e+02, -1.66893860e+02, 2.65522919e+01, 2.07407181e+02, 7.56258545e+01, -1.20566750e+02, 1.68667313e+02, -3.74505653e+01, -2.74624176e+01, 6.16770897e+01, -3.77194748e+01, -2.48264191e+02, 3.85292664e+01, -1.50520964e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.91080414e+02, 1.08105755e+01, 2.41298218e+02, -6.89335938e+01, -3.84979706e+01, 1.15534111e+02, 1.82165203e+01, -1.86847019e+01, 8.32165985e+01, -1.02730637e+02, 6.47277679e+01, -4.76232948e+01, 8.02088394e+01, -8.98775406e+01, -4.23134766e+01, 1.45877518e+02, 1.68536472e+01, 5.92792463e+00, 4.99991150e+01, 1.31623215e+02, 5.83446236e+01, -2.10806770e+01, 5.02811623e+01, -6.79594116e+01, 2.61618938e+01, 4.08175163e+01, 5.69347153e+01, 7.37486877e+01, 7.44672775e+01, 5.63325405e+00, -2.29764893e+02, 5.50410156e+01, 2.26119278e+02, 8.62248535e+01, 2.08033524e+01, 2.34858398e+01, -1.62039993e+02, 1.08243584e+02, 1.46900238e+02, 1.34521683e+02, 2.81825294e+01, -1.16488609e+02, -5.85900078e+01, 1.73250793e+02, 9.76689835e+01, -2.03759018e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.16514526e+02, -1.17052216e+02, -4.91777277e+00, -8.68883286e+01, 1.86644592e+01, 3.90022507e+01, 3.18189812e+01, -1.20419418e+02, -1.12250420e+02, 5.07412910e+01, -5.03331413e+01, 6.97251892e+01, -2.20708218e+01, 7.37625122e+01, -5.27069168e+01, 4.74246140e+01, 1.24728096e+02, 5.38428421e+01, -8.49598160e+01, -1.19387884e+01, -8.72923355e+01, -1.18827194e+02, 2.65218639e+01, -1.56816635e+02, 7.80680084e+01, 1.69501171e+01, 5.31437378e+01, 7.25622177e+00, 2.71921444e+01, 1.54061005e+02, -1.81372238e+02, -1.23805336e+02, -6.18844566e+01, -2.02820984e+02, -2.29358841e+02, -6.62072067e+01, 2.13282440e+02, -8.35622120e+00, -3.70531235e+01, 2.46184525e+02, 2.38832809e+02, 2.20934570e+02, -2.19295502e+02, -4.63693428e+01, -2.41744202e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.74298157e+02, -6.03627319e+01, 3.89796638e+01, 5.55433311e+01, 6.34318161e+01, 3.15922928e+01, -5.79880562e+01, -9.95403290e+01, -5.30535240e+01, 6.64405489e+00, -2.22661877e+01, -2.78855133e+01, -4.15738449e+01, -1.77600922e+02, 5.60068226e+00, -3.98287010e+01, 4.14669657e+00, -7.33472013e+00, -1.08208208e+01, 2.74775677e+01, -3.49522820e+01, -3.76124725e+01, -4.77917976e+01, 1.36276047e+02, -8.26806183e+01, -5.05860672e+01, 9.40767593e+01, 2.14245129e+01, -1.85702026e+02, -8.58719406e+01, 9.61184387e+01, -1.27735672e+02, 1.23629959e+02, 4.97750664e+01, -1.63868790e+01, -8.10352249e+01, 2.33143265e+02, -2.24786560e+02, 1.17590240e+02, 2.33220947e+02, -1.75294907e+02, 2.28461227e+01, 1.52759460e+02, -5.48205912e-01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.97965942e+02, 4.09950294e+01, -1.02771843e+02, 4.68635101e+01, -1.18423294e+02, -1.18479599e+02, -1.52255508e+02, -3.63859677e+00, 9.56049633e+00, 6.58642044e+01, -4.16359177e+01, -2.43397236e+01, -2.22847137e+01, 2.28402710e+01, -7.47503357e+01, -8.55913391e+01, 2.99121265e+01, -1.70901382e+00, 4.11776695e+01, -8.64484253e+01, -3.62525597e+01, 4.39127922e+01, -2.05453548e+01, 8.11930847e+01, -9.78181000e+01, -6.94092331e+01, -2.69759750e+01, -5.25708885e+01, -1.05705166e+01, 1.04351006e+02, 4.47004414e+00, 1.45278503e+02, -3.45585518e+01, -1.67195450e+02, -5.59032364e+01, -6.65321274e+01, 8.77674866e+01, 5.00622864e+01, 1.70615311e+02, 2.18079147e+01, -2.13667297e+02, -2.24742813e+02, 4.62639313e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.97489624e+02, -1.44419266e+02, 1.98107767e+00, -6.04104652e+01, -1.03747475e+02, 1.57748459e+02, 4.85732651e+01, -3.79861498e+00, -3.61019669e+01, -1.25487740e+02, -5.78528862e+01, -1.17231682e+02, 1.53545135e+02, 5.95123863e+01, -1.09437271e+02, 5.51984673e+01, -2.61053295e+01, -7.72231245e+00, 6.66603851e+01, -7.31296082e+01, -2.69217033e+01, -7.86112213e+01, -4.96519585e+01, -4.20251846e+01, -2.18847904e+01, -3.49032688e+00, 8.30606613e+01, -9.54951553e+01, 1.89176620e+02, -1.68215012e+02, -2.35896408e+02, -9.17342529e+01, 1.99122665e+02, -1.68170227e+02, -1.72329845e+01, -1.52848419e+02, -1.21303307e+02, 1.56255188e+02, -2.12877167e+02, -7.49517136e+01, -8.24503632e+01, 1.55311859e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.07309082e+02, 8.14337769e+01, -8.18889694e+01, 6.25668182e+01, -2.53470860e+01, 9.92369385e+01, 5.98138580e+01, 6.03740616e+01, 3.40139999e+01, 4.08447609e+01, -1.23370865e+02, 1.56589174e+01, 1.25197916e+01, 5.13593369e+01, 5.38004417e+01, -2.04366183e+00, 3.46429038e+00, 1.64747726e+02, 7.27671356e+01, -1.22866488e+01, 7.94410172e+01, 5.07296677e+01, 6.63664398e+01, 1.75694351e+02, 1.26274405e+01, -1.10016842e+01, 8.38173752e+01, 1.21311752e+02, 8.24246750e+01, 1.95906448e+02, -2.57638149e+01, -9.50969162e+01, 1.19683525e+02, -1.95070190e+01, 9.68757248e+01, -1.64210770e+02, 1.44376343e+02, 5.16190338e+01, -1.58581451e+02, 7.05704269e+01, -1.27641220e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.39234222e+02, 1.31592300e+02, -2.55419426e+01, -5.38301134e+00, 7.50016251e+01, 1.07231026e+02, 4.61620789e+01, 4.15447693e+01, -2.52683411e+01, 5.62453508e+00, 2.86715698e+01, 8.45602951e+01, -4.79163628e+01, 9.29349518e+01, 1.09565704e+02, 1.60865387e+02, 4.95661402e+00, 1.21165894e+02, -2.77504635e+01, -6.39731331e+01, -1.25843102e+02, -5.51387100e+01, 4.54183350e+01, 3.56040955e+01, 8.85337143e+01, 3.28900986e+01, -1.77332901e+02, 9.92222137e+01, -8.24153709e+00, 1.64200027e+02, 9.83826904e+01, -1.49088726e+01, 1.21555046e+02, 3.91915512e+01, -1.33484530e+00, -9.51596222e+01, -7.70338974e+01, 7.13173447e+01, 1.51688065e+02, 1.41637329e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.02986633e+02, 1.75012665e+01, -5.59001064e+00, -7.87838669e+01, 8.90889206e+01, 8.20969296e+00, 1.87121811e+01, -1.98615509e+02, 5.80739555e+01, 2.99569905e-01, -6.11247978e+01, 3.36530781e+00, -5.75944481e+01, 4.15922966e+01, 6.07245140e+01, 1.41891632e+02, 7.34960251e+01, -1.45583839e+01, -3.93020463e+00, 1.57058121e+02, -5.56933784e+01, -1.85531982e+02, 1.29761963e+01, 3.82205658e+01, 4.64309502e+01, 2.00683258e+02, -1.02196419e+02, -6.54013138e+01, -4.39192123e+01, 1.45777496e+02, 3.07692051e+01, -1.91614319e+02, 9.27744675e+01, 1.34954971e+02, -5.53055038e+01, -3.99076958e+01, -1.48942429e+02, 1.10510094e+02, -1.64988571e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.12707794e+02, -1.79511551e+02, 8.35545349e+01, 3.51165695e+01, -2.61699123e+01, 7.80633545e+01, 5.64615784e+01, -2.76772690e+01, 7.17738266e+01, -1.98051014e+01, 5.27134933e+01, -8.31406326e+01, -9.12008514e+01, 6.52064657e+00, 1.42686357e+01, 1.05551491e+01, 3.35673857e+00, -7.70260620e+01, 3.72617188e+01, 1.04693634e+02, -3.63387871e+01, 6.54018631e+01, -1.62921280e+02, -1.01522551e+01, -1.17552170e+02, -5.49479332e+01, -4.52354393e+01, 2.18408108e+01, 1.99174774e+02, 5.58371201e+01, -1.30782379e+02, 1.00073143e+02, 1.16352623e+02, -9.15323486e+01, 4.52650948e+01, -1.77718536e+02, -3.07998543e+01, 1.66133194e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.82679230e+02, 1.75287418e+01, -2.41694050e+01, 1.15093971e+02, 3.14991512e+01, 4.09192848e+01, -3.83480835e+01, 7.04006348e+01, 5.09637184e+01, -9.47815094e+01, 8.61740036e+01, 3.34271812e+01, -1.36822195e+01, 7.85643845e+01, -1.08383881e+02, 1.49503372e+02, 3.57298584e+01, 6.50731888e+01, 1.05469917e+02, 8.73189697e+01, -2.17338104e+01, -3.94213715e+01, 8.97312393e+01, -3.82087212e+01, -1.96499603e+02, -5.43299370e+01, -5.52835503e+01, -1.29212769e+02, 1.44204788e+02, 1.91481155e+02, -1.88283783e+02, -1.41424236e+01, 1.48895370e+02, 1.03787987e+02, -4.00807381e+01, 9.39138870e+01, -1.11107262e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.08214325e+02, -1.01204086e+02, 3.44344711e+01, 9.78280411e+01, -2.86712627e+01, -5.56974068e+01, -7.00710144e+01, 5.03080215e+01, 8.33946800e+00, 2.68616505e+01, 7.40704679e+00, -1.82006276e+00, 5.54016609e+01, 9.61355515e+01, -5.73444405e+01, 2.18658600e+01, -4.35028076e+01, -2.16856632e+01, 4.47589397e+00, 6.38850250e+01, 3.80345268e+01, 1.82031860e+01, -1.39532410e+02, 7.95302048e+01, -9.95157089e+01, 2.33040054e+02, 1.21432854e+02, 3.75891380e+01, 1.19106827e+02, 2.48378968e+00, 2.14874481e+02, -1.28095825e+02, -7.86705933e+01, -1.37993622e+02, -2.40044464e+02, -1.48598007e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.89783783e+02, 2.61773872e+01, 1.48605545e+02, -3.20931854e+01, -8.60211029e+01, 3.20089264e+01, -4.01425247e+01, -3.05804729e+01, -4.91096458e+01, 6.94688721e+01, 1.30594696e+02, 8.89653549e+01, 1.38680161e+02, 3.21516633e+00, 1.92086697e+01, 1.24859749e+02, 6.86144104e+01, -2.44163818e+02, 2.10376568e+01, -1.00767326e+02, 3.08317547e+01, -8.76789322e+01, 1.32893753e+02, -6.44752274e+01, 1.70783234e+02, -7.24740906e+01, 1.25429039e+02, 3.78939781e+01, 9.42913818e+00, 4.35326080e+01, -1.00363106e+02, -3.02696762e+01, -4.27504768e+01, -2.04189392e+02, 7.68877869e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.25470459e+02, -1.31756487e+01, -1.99992275e+01, 3.34166183e+01, -8.78470383e+01, 1.28637970e+02, -4.59645958e+01, -1.32765488e+02, 1.46861877e+01, 9.60403538e+00, -4.43682137e+01, 5.90441933e+01, -1.03707001e+02, -4.03577118e+01, 5.10834961e+01, -9.02098618e+01, -1.11082504e+02, -1.03443687e+02, 1.94711639e+02, -8.17075348e+00, 1.37483200e+02, 3.05449486e+01, 1.86740097e+02, -7.87500916e+01, 1.17675804e+02, 6.00887108e+01, 1.28512665e+02, -6.66160736e+01, 2.13762627e+01, 1.34338058e+02, -1.99562241e+02, 2.09346573e+02, 1.24279671e+02, -1.75654892e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.70407410e+02, -1.21266273e+02, -1.77701645e+01, -6.19785690e+01, 2.64975395e+01, -3.26965561e+01, -6.47338333e+01, -5.66444702e+01, 1.23561020e+02, 9.14225159e+01, 1.65724659e+01, -7.87227478e+01, 3.24505157e+01, 1.51274796e+02, 1.69748898e+01, -1.03938881e+02, -5.12681007e+01, 7.82763824e+01, 4.19464035e+01, -7.76790848e+01, -9.49721603e+01, -1.56332407e+01, -1.40154999e+02, 4.83104286e+01, -2.20852222e+01, 8.77063370e+01, 1.36793976e+02, 1.55199919e+01, 1.26068512e+02, 6.86401978e+01, -7.33807297e+01, 3.89638672e+01, -2.95447578e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.65091064e+02, 3.67449455e+01, 2.61178379e+01, -1.22459305e+02, -7.42200928e+01, -5.93329239e+01, 2.12636993e+02, 6.60040741e+01, 9.33437920e+00, -5.95518303e+01, 5.52984810e+01, 2.63704128e+01, 6.15807076e+01, 2.06003067e+02, -1.25161873e+02, 9.58113556e+01, -2.90824795e+01, -1.76083794e+01, -1.59460236e+02, 1.31249008e+02, -1.92135849e+02, 1.72366180e+02, -1.16450851e+02, 2.66243305e+01, 8.34304733e+01, 8.34782028e+01, -1.86592346e+02, 1.25785828e+02, 4.14071426e+01, 6.32625465e+01, 7.26284256e+01, 3.35928802e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.20508514e+02, -1.35046280e+02, -8.23201981e+01, -3.26098824e+01, -8.75222244e+01, -8.59274826e+01, -6.97950287e+01, 3.68370094e+01, -8.54520035e+01, 5.14520502e+00, 1.12844610e+01, 2.53412952e+01, 6.09242249e+01, 2.08137379e+01, -7.25340366e+00, -1.67897827e+02, -4.88750000e+01, 1.61064728e+02, -1.61257980e+02, 4.49209061e+01, 1.87407272e+02, 1.54247482e+02, 3.07405605e+01, -9.17294846e+01, 5.26410294e+01, 1.68591263e+02, 1.75807400e+01, 1.76619843e+02, 1.02637024e+02, -1.72852875e+02, 1.31917130e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.68100250e+02, 1.43981590e+01, 2.33385868e+01, 1.04547432e+02, 6.89168320e+01, -3.52091968e-01, 1.02721283e+02, 1.17554855e+02, 5.29525337e+01, -5.99782562e+00, 2.40405102e+01, 6.49776611e+01, -1.11157249e+02, 8.90245056e+01, -1.64608749e+02, -7.64628077e+00, -8.75003815e+01, 7.25024185e+01, -4.84131927e+01, 1.31288803e+02, -4.61800146e+00, 4.83056755e+01, -1.51668900e+02, 5.29789543e+01, 9.94529724e+01, -6.29755096e+01, -5.35252953e+01, -1.38016998e+02, 6.95257645e+01, -1.15730392e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.29622833e+02, -4.97540092e+01, -2.37363949e+01, -5.57590599e+01, -7.12380066e+01, 1.55203557e+00, -5.58621292e+01, -2.66103954e+01, -9.61696243e+01, 2.27196884e+01, -1.54174179e+02, -8.66771927e+01, -1.16844711e+02, -2.12551025e+02, 7.64079380e+00, 5.45665894e+01, -5.17064285e+00, 1.07765770e+02, -1.65983154e+02, 1.84517136e+02, 2.30886440e+01, 8.48210602e+01, -4.48635788e+01, -1.31633434e+01, 1.76291611e+02, -1.18921709e+01, -1.45997955e+02, -2.34204369e+01, -5.09206123e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.02473816e+02, 2.24857529e+02, -8.37415886e+00, 1.32673416e+02, 4.16868782e+01, -9.17534790e+01, 8.92105560e+01, -7.41908340e+01, 1.84777176e+02, 2.01072037e+02, -1.38967804e+02, 3.76796570e+01, 2.17265987e+00, 6.81923065e+01, 2.36095352e+02, 5.08503876e+01, 3.87777495e+00, 6.65236130e+01, 2.05199982e+02, 6.05403099e+01, 6.78475647e+01, -1.07367027e+02, 2.40219208e+02, 1.23833870e+02, 1.20055832e+02, -3.13641052e+01, -1.84036179e+01, -3.58690147e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.29618683e+02, -3.33187943e+01, -1.28062271e+02, -5.45585060e+01, -1.39008301e+02, 3.48511391e+01, -1.35580139e+02, -5.40202332e+01, 3.63220482e+01, 1.43339417e+02, -6.38635025e+01, -8.09085655e+00, 9.81891174e+01, 3.20184593e+01, -1.17975739e+02, 8.39138260e+01, 1.37123154e+02, -3.96665115e+01, 7.11257935e+01, 1.52831650e+02, 2.03769760e+01, 1.28876099e+02, 4.22312241e+01, -1.17812286e+02, 8.31572495e+01, -3.36340370e+01, 1.17952553e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.61064697e+02, 3.82435455e+01, 1.67897934e+02, 4.97332497e+01, -9.60541534e+01, 1.54167414e+01, 2.92999401e+01, -1.32586243e+02, -1.41767487e+02, 2.99079857e+01, 9.80248413e+01, 1.52770004e+02, -9.26377106e+00, 7.83538437e+00, 8.58400440e+00, -1.01180954e+02, 1.36225769e+02, 1.49401886e+02, 6.75390396e+01, -9.11664963e+01, -2.09315933e+02, -1.43275955e+02, 1.85428894e+02, 1.29181992e+02, -1.18530869e+02, -1.66010040e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.12013336e+02, 1.08837652e+01, -5.99310341e+01, 2.04628277e+01, 2.19348602e+01, -6.39389114e+01, -3.64968872e+01, 1.46479446e+02, -1.87228050e+01, 2.78994789e+01, -7.23590698e+01, -2.62453079e+01, 1.57968407e+01, 5.93414497e+01, -8.43399143e+00, -9.14365005e+01, -1.47399185e+02, 1.18912170e+02, 1.91980171e+01, 1.16440308e+02, 8.98238831e+01, 9.39350815e+01, 1.77401764e+02, -8.65982342e+00, -7.85976486e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.02492462e+02, -7.35858688e+01, 5.54313126e+01, 6.86667175e+01, -1.25962076e+01, -7.64957857e+00, 4.17153893e+01, -4.33914566e+01, 4.83947182e+01, 3.41431847e+01, -8.31776714e+00, -1.30623589e+01, 1.25487480e+02, 1.83012985e+02, 1.98702759e+02, -1.45371313e+01, -1.12969027e+01, -4.05839844e+01, 1.66520462e+02, -8.63990326e+01, -1.30936676e+02, 5.37832832e+01, 1.01061516e+02, -1.99049316e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.61118713e+02, -2.39957428e+01, -8.40626221e+01, -1.68464691e+02, 5.15807571e+01, -1.63263489e+02, 8.86732712e+01, -1.26310930e+01, -7.02703171e+01, -1.60249023e+02, 1.42997086e+02, -1.25324280e+02, -1.92763481e+01, 1.49768438e+01, -5.03480530e+01, 1.37508774e+02, 1.44978821e+02, 9.57762299e+01, 1.51868683e+02, 9.07083359e+01, 1.06142105e+02, -1.33404922e+02, -4.72282839e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.50716736e+02, -8.03975983e+01, 2.03228474e+01, -6.97260513e+01, 5.36292601e+00, -2.62696247e+01, -1.69708004e+01, -1.41587038e+01, 5.16692972e+00, -1.44132586e+01, 6.31512222e+01, 1.03742584e+02, -1.22047997e+02, 5.77113152e+01, -7.13928146e+01, -6.28338203e+01, 2.46968632e+01, 1.34207642e+02, 4.92350235e+01, 5.34422607e+01, -1.64065109e+02, 1.21327339e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.19130005e+02, 1.01298809e+01, -9.09828262e+01, 4.64982910e+01, 4.13495064e+00, -1.34872070e+02, -1.94799862e+01, -5.91307945e+01, 1.02488014e+02, -1.34708206e+02, -1.30293106e+02, -2.32203274e+01, -8.95526171e-01, 1.26339684e+01, 2.93568554e+01, -9.75342636e+01, -7.84858780e+01, 2.40679665e+01, -1.88981247e+01, 7.73498383e+01, 7.29062271e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.77862457e+02, -2.14183483e+01, 1.05928505e+02, 1.03438438e+02, -8.75316315e+01, -8.45970688e+01, -7.97552567e+01, -7.63711700e+01, 1.08213364e+02, -7.86264496e+01, 2.44560013e+01, -9.99846497e+01, -2.99024982e+01, 9.69001465e+01, 7.88930893e+01, -2.57592907e+01, -5.74590034e+01, 7.49051895e+01, 2.27268600e+01, -6.66452103e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.53487076e+02, -3.78693008e+00, -2.59061241e+01, -6.50942001e+01, 4.41557770e+01, -7.77448130e+00, 2.57557750e+00, -8.71569901e+01, -6.04069290e+01, -2.53041420e+01, 5.31526527e+01, 1.21848984e+02, -7.71827011e+01, 2.58078346e+01, -1.19881683e+02, 1.20845520e+02, 8.96611786e+01, -2.50661349e+00, 1.00783379e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.29936066e+02, -5.24691200e+00, 2.87416515e+01, 1.17859177e+02, -1.56266357e+02, -2.72667007e+01, -1.30282181e+02, -1.01315025e+02, -6.89075394e+01, 2.84362774e+01, -3.69533043e+01, -6.28071556e+01, -1.47241074e+02, 8.09635391e+01, 1.36727341e+02, -7.31809235e+01, 9.05354738e-01, 3.34533043e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.58497650e+02, -3.15342641e+00, 5.20423508e+01, 7.24591904e+01, 1.49625778e+01, -1.74532471e+02, 1.00044754e+02, -6.92515488e+01, 1.60980423e+02, -7.49604034e+01, -4.63602219e+01, 8.71624069e+01, -3.83424683e+01, 1.61330322e+02, 5.64889488e+01, -4.10712471e+01, 8.25534439e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.13831360e+02, 1.08122005e+01, -1.65988159e+01, -7.72937698e+01, -1.10133957e+02, -1.51749481e+02, 5.82981720e+01, 4.24005089e+01, -9.47852325e+01, 8.72929535e+01, -1.14881142e+02, 1.15125519e+02, -7.88253098e+01, 4.46657066e+01, 1.39003662e+02, -1.82181778e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.89282745e+02, -1.24558098e+02, 2.86739864e+01, 9.37672577e+01, 1.44127579e+02, 4.42492723e+00, -1.04416504e+02, 1.06790199e+02, -7.69855070e+00, 1.39556793e+02, 1.27496277e+02, 8.62635956e+01, 2.07790432e+01, 3.24385910e+01, -7.92296219e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.56155701e+02, -8.80482712e+01, 1.17872864e+02, 6.28091469e+01, -9.33015976e+01, 2.58868103e+01, 1.83438663e+01, 1.70502224e+01, 9.84515610e+01, -5.94389420e+01, -1.94004555e+01, -9.29320755e+01, -8.31301880e+01, -7.23417435e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.20619217e+02, 8.03923950e+01, -1.04501732e+02, 6.44229736e+01, -3.29314003e+01, 2.43011112e+01, 7.49697723e+01, 1.91958790e+01, 1.27880659e+01, 4.52517080e+00, -8.53084412e+01, 1.00002136e+02, -2.21724854e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.15662170e+02, 4.02118759e+01, 1.03015488e+02, 1.12457132e+01, -5.00677338e+01, 2.41788921e+01, 9.54426575e+01, 2.01820602e+01, 1.88162842e+01, 1.02888519e+02, 4.53607292e+01, -5.71925926e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.93809738e+02, -8.42976379e+01, 1.33865242e+01, -2.59627934e+01, -1.50120821e+01, 2.36200428e+01, 2.49448452e+01, 9.60762863e+01, -9.13225479e+01, -2.45612183e+01, -4.54805517e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.10994705e+02, 6.31055298e+01, -1.22840395e+01, -1.02811691e+02, -1.90140705e+01, -1.79196835e+01, 4.04523926e+01, -5.54266024e+00, 4.53170395e+01, -1.32757342e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.82826782e+02, -8.02260971e+01, -6.07852316e+00, 6.36038094e+01, -8.36942673e+01, 6.72870712e+01, 9.10806732e+01, -2.69303703e+01, -6.77709351e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.00352097e+02, -8.81713486e+01, -6.26500778e+01, 4.47206802e+01, -6.63242111e+01, 6.11492577e+01, -2.15102901e+01, -4.97934990e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.92574875e+02, -1.89510098e+01, 1.78880620e+00, -7.88275452e+01, 1.70102768e+01, -8.48473511e+01, -4.28930740e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.69149307e+02, -7.01327057e+01, 7.46393509e+01, -1.34721985e+01, -7.32466202e+01, 6.86263275e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.44616959e+02, 3.78329239e+01, 4.91787033e+01, -8.84605503e+00, 2.91303673e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.50519653e+02, 6.71853714e+01, 3.19158974e+01, -7.36191330e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.46282089e+02, -3.88422394e+01, 1.93981018e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.53404739e+02, -7.05484695e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.36881912e+02} } };
-
-const std::array<float, 60> test_solution_big = {  0, -0, -1, 0, -0, -0, 1, 0, 0, -0, 0, -0, -0, -0, 0, 1, 0, 0, 0, -0, -0, -0, -0, -0, 0, -0, 0, 0, 0, -0, 0, -0, 0, 0, 0, 0, 0, 0, 0, 0, 0, -0, 0, -0, -0, -1, -0, -0, -0, 0, 0, 1, 0, 0, -0, -0, -0, 0, -0, 0 };
-
-const std::array<std::array<float, 20>, 20> test_mu_small = { { { 310.95657, -152.87987, 81.47761, 61.63883, -17.108498, -15.243285, 52.370655, 108.87372, 114.97746, -100.52528, -113.50459, 32.943504, 48.926445, 21.539984, 45.427563, -79.80857, -35.979298, -4.4700775, 95.74005, -130.8382},
- { 0., -234.14899, -68.073265, -97.029396, 67.69854, -45.849476, 24.222107, -40.835533, 100.01009, -105.815994, 64.90127, -93.899185, 98.074875, -30.032288, 40.8844, 32.77453, 10.884139, 62.86333, -91.9531, 36.816414},
- { 0., 0., -235.68079, 98.01093, 101.970894, 99.58868, 106.60249, -31.247442, 68.24517, -83.04165, -97.47559, -55.455597, 15.233033, -95.41965, -99.55332, -104.173454, -42.219933, 75.392136, -30.52341, -7.067175},
- { 0., 0., 0., 373.61185, 52.704525, 109.29272, -55.45065, -45.106987, 43.154602, -180.00504, 20.634502, -63.151974, 148.01558, 19.430471, 174.17596, 24.021017, 2.2237465, 100.170456, -180.13036, 151.96693},
- { 0., 0., 0., 0., -356.76935, 161.67087, -82.84156, -93.183014, 172.01004, -96.05831, -96.36346, -105.35294, 119.418564, -115.94216, 154.4217, 144.27682, -100.111206, 26.16108, -82.51057, -134.36534},
- { 0., 0., 0., 0., 0., -330.04272, -159.57295, -151.98651, 129.4496, 32.925636, -37.70098, 5.319791, 89.625015, -13.916242, -30.06193, -42.61862, -42.616432, -4.9494314, -28.186255, -14.948075},
- { 0., 0., 0., 0., 0., 0., 249.32005, 123.24647, -57.046436, 86.93297, 1.1256452, -49.681755, -115.09167, -76.402596, 40.180843, 78.209625, 112.77826, -50.94506, 112.753815, 39.90859},
- { 0., 0., 0., 0., 0., 0., 0., -241.48967, -106.72415, -88.36093, -85.3396, -66.08973, 72.765495, 17.753405, 53.73414, 46.32324, 87.68742, -63.117607, 85.09047, 30.696415},
- { 0., 0., 0., 0., 0., 0., 0., 0., -191.17126, 34.5826, -52.809868, 80.39618, 94.17156, 90.388985, 22.990557, 70.03452, 60.15338, 15.50781, 39.882694, 6.0649223},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 169.99902, 75.17026, -22.079172, -23.528748, 20.505486, 24.132507, 0.91698325, 45.61528, 84.293365, -56.143097, 23.992374},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., -128.91632, -31.141186, -35.430202, 43.541035, 56.071404, 29.513527, 34.08576, 2.012616, 1.2261888, 22.051527},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 155.20854, 65.57702, 24.910547, -16.293734, 48.008087, -14.684272, 16.616346, 15.026151, -35.878014},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., -168.71866, -68.42938, 27.298466, 55.758804, -3.2235277, -35.998623, -43.36375, 31.927486},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 153.21968, -3.7473776, -67.885765, 41.157864, -62.210476, -47.048443, -61.933846},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., -144.52457, -2.1524777, -30.090687, -1.0992928, 68.44065, 52.487213},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., -159.35155, 36.52487, 20.918535, -25.088203, 18.503382},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 168.7975, 12.329721, 84.31132, 18.75195},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 157.65189, -72.59766, -31.100744},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 123.29349, 19.945765},
- { 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 167.34836 } } };
-
-const std::array<float, 20> test_solution_small = { { 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, -1, 0, 0, -1, -1, 0, 1, 0, -1, 0 } };
-
-const std::array<std::array<float, 50>, 50> test_mu_normal = { {{ 6.98995728e+02, 3.01019012e+02, 2.71596802e+02, -3.47437042e+02, 3.30056396e+02, -3.34421234e+02, -2.35323334e+02, 3.24099274e+02, -2.58682556e+02, -2.64706909e+02, -2.54901428e+02, 3.25029175e+02, -2.01818115e+02, 3.06993286e+02, 1.21571564e+02, 2.81923035e+02, 3.23492676e+02, -2.89398071e+02, -3.27299591e+02, -3.39424103e+02, -2.85053253e+02, -5.99731865e+01, -2.88413787e+00, -2.16672287e+02, 1.52616104e+02, -3.45228149e+02, -3.19260895e+02, 3.27652954e+02, 1.00228081e+02, -5.65668716e+01, 2.56739769e+01, 3.43923431e+02, -3.05335205e+02, 6.12564545e+01, 2.30098694e+02, -2.04181519e+02, 1.54235563e+01, -7.05669556e+01, -1.79733582e+02, 1.70491745e+02, 3.15326691e+02, -6.22679062e+01, 1.48052414e+02, -1.03319664e+02, -3.08001892e+02, -1.22324074e+02, -8.93882446e+01, -2.27041458e+02, -1.43890732e+02, 6.35569000e+01},
- { 0.00000000e+00, 7.08164917e+02, 2.78942352e+02, 1.90913376e+02, 2.68362274e+02, 2.81448761e+02, -4.61505508e+01, 1.68954926e+02, 2.72650299e+02, 1.19192314e+02, 2.38975662e+02, 9.71483231e+01, 1.31109421e+02, 1.81654266e+02, -5.31999702e+01, 2.00236969e+02, 2.31085373e+02, 2.21196106e+02, 2.99302307e+02, 1.29101440e+02, 1.92143738e+02, -4.08717384e+01, 7.78253479e+01, 2.23012283e+02, -8.78444366e+01, 3.37453041e+01, 2.51497360e+02, 1.96945984e+02, -1.03071411e+02, 2.86884735e+02, -3.30307739e+02, 2.11036301e+02, 2.48183289e+02, -7.84582214e+01, 1.67919815e+02, 6.58582687e+01, -1.35558517e+02, 3.12439941e+02, 1.68918967e+00, 2.25997833e+02, 4.40062332e+01, -1.14481468e+01, 3.23545258e+02, 1.86034607e+02, -2.76779785e+01, 2.05182251e+02, -2.72908813e+02, 1.45969940e+02, -3.39158508e+02, 3.03529816e+02},
- { 0.00000000e+00, 0.00000000e+00, -6.50390137e+02, -2.72597870e+02, -2.60134827e+02, -2.75228271e+02, -2.73573364e+02, -2.04418823e+02, -1.10557701e+02, -2.40444550e+02, -2.85885284e+02, -5.27503853e+01, -2.37771423e+02, -2.04499603e+02, 3.21864838e+02, -1.74278519e+02, -8.54116898e+01, -1.49068405e+02, -2.01765701e+02, -6.57799683e+01, -1.60183121e+02, -7.32889099e+01, 1.30625381e+02, -1.77656326e+02, 3.20019470e+02, -1.17066460e+02, -9.26720276e+01, -2.39943573e+02, 4.10225220e+01, -1.12844696e+02, -1.15067635e+02, -1.33554504e+02, -1.81619034e+02, -5.49704895e+01, -1.75171753e+02, 1.28013336e+02, 8.81239929e+01, -1.82253723e+02, 3.24030304e+02, -6.13149109e+01, 2.41407757e+01, 8.01316910e+01, -4.04980898e+00, -2.81045715e+02, 1.49702621e+02, -1.52472794e+02, 9.96519012e+01, -3.21218781e+02, 2.30824631e+02, 3.89353600e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.78972351e+02, 4.88981323e+01, 2.09287552e+02, 1.27325951e+02, 1.33303772e+02, 2.37249176e+02, 1.96793396e+02, 6.11591110e+01, 1.26666870e+02, 1.56959808e+02, 2.05679001e+02, 1.48409012e+02, 1.16811523e+02, 1.94776154e+02, 1.41340729e+02, 7.69938126e+01, 1.65220016e+02, 1.07270332e+02, -7.44100189e+01, -2.52104874e+02, 6.67825623e+01, 1.24468445e+02, 1.12684921e+02, 3.97489281e+01, 2.57126343e+02, -1.25608223e+02, 1.99529770e+02, 2.24256073e+02, 1.73611771e+02, 2.41620804e+02, 3.88578949e+01, 1.57342957e+02, -1.90130463e+02, 1.41462814e+02, 2.87059448e+02, -1.58429337e+01, -1.34583588e+02, -2.39958755e+02, -1.40280361e+01, -2.06740784e+02, -1.16897306e+01, -2.10397873e+02, 1.38247177e+02, -2.45657593e+02, -1.10650841e+02, -2.45118393e+02, 8.85272217e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.05613800e+02, -5.02980728e+01, -6.64955673e+01, -2.03529007e+02, -1.68228851e+02, -1.07976990e+02, -1.23193832e+02, -1.09354431e+02, -1.86549240e+02, -6.85126801e+01, 5.13915405e+01, -1.06011086e+02, 1.73436127e+01, -1.65804031e+02, -7.99873886e+01, -1.37886917e+02, -7.32831802e+01, 4.45745773e+01, 1.29701364e+00, -6.96516724e+01, -1.33970139e+02, -5.36370773e+01, -8.97776642e+01, -1.09203959e+01, -3.22419090e+01, -5.44659767e+01, -5.87137375e+01, -1.65231293e+02, -8.02026367e+01, -7.66085205e+01, 6.80874176e+01, -6.06318321e+01, 8.65937576e+01, -4.31530342e+01, 9.87628021e+01, -2.28843750e+02, 2.38393036e+02, 6.68253937e+01, -2.33970581e+02, -2.36254379e+02, 2.52191208e+02, -2.12642622e+00, 1.08893951e+02, -1.87243866e+02, -7.14671173e+01, 1.52442141e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.91241272e+02, -2.46506047e+00, -7.05714264e+01, -1.80864365e+02, -2.55515442e+02, -1.10684395e+02, -2.55232285e+02, -6.42748337e+01, -2.63657513e+01, -1.82396286e+02, -2.08672638e+02, -1.77973728e+01, 2.76393757e+01, -1.96344757e+02, -9.54373245e+01, -1.71845673e+02, 2.09763412e+01, -2.11820602e+02, -1.32249741e+02, -1.99679962e+02, -1.40237854e+02, -9.72482758e+01, -1.81338867e+02, 6.39944992e+01, -5.85206146e+01, -1.76921967e+02, -8.96382141e+01, 4.81942253e+01, -3.48812943e+01, -2.88757904e+02, -2.51286255e+02, 2.54826508e+02, -2.75897888e+02, 2.21920456e+02, 1.13975601e+02, -6.00545845e+01, -2.12332020e+01, -8.65718155e+01, 3.29372215e+01, 2.27561325e+02, -8.60067673e+01, -1.73756088e+02, 7.36419983e+01, 9.04678650e+01, -1.48309845e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.82565063e+02, -6.62675705e+01, -1.68274750e+02, -9.31090775e+01, -2.23643997e+02, -6.33604279e+01, -9.04279099e+01, -1.13941666e+02, 6.40504684e+01, -5.94566460e+01, -1.63124714e+01, -1.65373154e+02, -9.45231781e+01, -6.58572311e+01, -1.10267700e+02, 1.40901321e+02, -9.14647903e+01, -5.82121658e+01, -7.13210526e+01, -5.97973099e+01, -2.25209928e+00, -1.26657715e+02, 4.49205360e+01, -1.49989716e+02, -1.83975021e+02, -3.83528256e+00, -9.18682861e+01, -1.45918457e+02, 2.56304901e+02, -2.83319641e+02, -1.02400879e+02, 1.52153381e+02, 1.63216003e+02, 1.27010139e+02, 2.40414474e+02, 5.62635002e+01, 1.78031418e+02, -5.30501900e+01, -2.87125397e+01, 1.99144775e+02, 2.79688416e+02, 5.95454330e+01, -2.30501236e+02, 2.01360458e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.48354065e+02, 1.55444214e+02, -5.10100317e+00, 4.69662895e+01, 2.08840622e+02, 1.02373001e+02, 1.73966675e+02, -8.27329254e+01, 1.49983368e+02, 8.99784393e+01, 1.40492493e+02, 2.26583130e+02, 1.91603088e+02, 1.38689852e+01, -1.22382629e+02, 1.06057281e+02, 6.03512878e+01, 1.46941727e+02, 8.87104492e+01, 1.26935081e+02, 3.16652584e+01, 6.88792114e+01, 1.14292898e+01, 3.03761253e+01, 8.68133850e+01, 1.63953369e+02, -4.96841698e+01, 1.34290787e+02, 2.41930511e+02, 2.20226593e+02, -7.88237858e+00, 1.41759155e+02, -2.18821014e+02, -2.41903091e+02, -4.22527847e+01, -1.10877930e+02, 1.83447021e+02, -2.03896961e+01, 1.95754623e+02, 6.52769775e+01, -1.94762787e+02, 4.26777992e+01, 9.55812683e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.53260193e+02, -6.34955902e+01, -6.19112015e+01, -3.07165909e+01, -1.92529831e+02, -6.47805405e+01, -1.52601135e+02, -1.06537529e+02, -2.74829159e+01, -1.99506577e+02, -1.52787277e+02, -9.60634689e+01, -1.29597519e+02, 2.67467041e+02, -5.48391724e+01, 7.62330723e+00, -6.12534790e+01, -7.19925079e+01, -7.62908707e+01, -3.86052742e+01, -5.77665520e+01, -1.82842850e+02, -1.63639755e+02, -1.29973404e+02, -1.49621552e+02, -6.18106995e+01, 1.39302948e+02, 2.43128479e+02, -2.57213348e+02, 4.57999496e+01, -4.97234688e+01, -1.15106674e+02, 2.90615606e+00, 2.39371399e+02, -2.70482819e+02, -1.25287910e+02, 4.27454758e+01, 1.96844727e+02, 2.53444565e+02, 2.57762390e+02, 1.74008102e+02, 1.54894501e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.10122559e+02, -1.78639465e+02, 3.35559807e+01, -8.93768158e+01, -4.90212479e+01, -5.99675331e+01, -7.07767715e+01, -1.32387833e+02, -2.84292316e+00, -1.97330360e+01, -7.17490053e+00, -7.36312332e+01, -5.32743797e+01, 7.80136719e+01, -1.10132980e+02, 1.11444054e+02, 2.91400127e+01, -1.91941719e+01, -1.29474594e+02, 6.25045128e+01, 1.35123196e+01, -1.14297035e+02, -8.79352951e+01, -5.34774780e+00, -7.32826767e+01, -1.75124664e+02, 1.48157425e+01, 2.00601791e+02, -1.78360519e+02, -9.34070511e+01, -5.86838493e+01, 1.36529526e+02, 7.67620010e+01, -4.58051987e+01, -2.01063919e+02, 5.35873413e+01, 9.54568939e+01, -2.26661789e+02, -1.91185928e+02, -7.39709320e+01, -1.75676376e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.63259094e+02, -2.26433945e+02, -9.33935852e+01, 1.18183994e+01, -1.15669144e+02, -8.97467651e+01, 4.84375877e+01, -8.69978714e+01, -1.03219429e+02, -1.23428368e+02, -1.37135040e+02, -1.25117714e+02, -8.64857788e+01, -6.48665619e+01, 7.31338024e+00, -9.13577118e+01, -4.96265793e+00, -9.60613632e+01, -2.34072447e+00, -4.36555557e+01, -1.23638754e+01, -1.45998947e+02, -9.16018448e+01, 1.12249550e+02, -1.99771057e+02, -6.27481537e+01, -1.11994408e+02, -1.87293579e+02, -4.60876236e+01, 1.35697067e+02, -4.53687763e+00, 3.94289246e+01, 3.15442772e+01, 1.37368896e+02, -7.86230392e+01, -1.29720795e+02, 1.39105270e+02, -2.51058807e+02, -1.26169739e+01, -1.64908962e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.00185394e+02, -6.70274887e+01, -1.09234718e+02, -2.15303478e+01, -8.32470226e+00, -1.33055668e+01, -2.22912216e+01, -4.53527832e+01, -3.61327934e+01, -5.42746582e+01, 7.39990854e+00, -8.16320343e+01, 1.92418861e+01, 4.29322319e+01, -3.85810776e+01, 2.87117691e+01, 4.40052376e+01, -1.33336121e+02, -1.00712929e+01, -1.54923264e+02, -3.31515198e+01, -8.01333313e+01, 6.84864120e+01, 1.15737984e+02, -1.15404465e+02, 1.28041321e+02, 5.05765915e+01, -1.60592575e+02, 1.41107620e+02, -1.34209747e+02, -4.20944862e+01, 2.18250900e+02, -3.10156059e+01, 1.75943039e+02, -2.12328568e+02, -1.83572556e+02, 5.84264278e+00, 1.01773094e+02, -2.33118454e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.45193024e+02, -3.65584641e+01, 1.04885651e+02, -1.99562359e+01, -1.98493713e+02, -1.50797516e+02, -1.03698318e+02, -4.95123577e+00, -2.41289349e+01, -3.01953430e+01, -6.55425568e+01, 7.32700806e+01, -2.68472748e+01, -1.40478867e+02, 1.00681694e+02, 4.19611702e+01, -1.50271912e+02, -3.13899326e+01, -1.46077681e+01, -9.26623688e+01, 8.03375721e+00, 1.13612770e+02, 9.95002213e+01, 6.40410995e+01, -2.05945683e+01, 3.83902397e+01, 5.49339104e+01, 2.15772018e+02, 1.71485001e+02, 1.94826431e+02, 2.02419403e+02, -1.72407455e+02, -2.31188660e+01, 6.71411743e+01, 1.94695480e+02, -1.11632938e+01, 1.23912682e+02, 1.13661674e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.71225677e+02, -1.43230343e+00, -8.36307220e+01, 2.06751423e+01, -8.48707657e+01, 3.39922218e+01, 9.19754219e+00, 8.42824554e+00, 5.45601273e+01, 1.24749886e+02, -1.00752731e+02, 1.70992298e+01, -3.07287884e+01, -5.45468636e+01, 9.44463577e+01, -2.99242764e+01, 6.25285568e+01, 3.26834450e+01, 1.71100903e+01, -9.28219376e+01, 5.07964211e+01, -2.31615448e+02, -5.07310638e+01, -1.55300385e+02, 1.14774994e+02, 1.22904549e+02, 1.62717163e+02, 1.68593552e+02, 3.08184910e+01, -1.89582581e+02, 5.56150436e+00, 1.65433853e+02, 1.40357803e+02, -3.20379944e+01, -1.21335487e+02, -4.71101494e+01, -1.25058136e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.56368317e+02, 5.98137817e+01, 1.40173325e+02, 8.38768082e+01, 1.85711861e+01, 3.17582569e+01, -5.08643532e+01, -1.67710129e+02, -1.45919762e+01, -8.73366699e+01, -9.88742905e+01, -1.13263405e+02, 1.11084908e+02, 4.74255524e+01, -1.57581284e+02, 2.79570484e+01, -7.25879669e+01, 5.15629044e+01, 4.09389915e+01, -6.33444824e+01, -1.96892197e+02, -1.82591187e+02, 8.46672516e+01, 1.97229187e+02, 3.08126450e+01, -1.25822617e+02, 2.05945847e+02, -1.26704283e+01, -7.15130157e+01, -6.15416946e+01, 1.59768127e+02, 8.50605545e+01, 1.67502609e+02, 1.66230713e+02, -1.51450653e+02, 1.98780701e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.91219391e+02, 4.86569443e+01, 9.96937561e+01, 5.20334320e+01, -8.51903076e+01, -4.72002554e+00, -1.28991976e+01, 3.89507217e+01, 2.24916718e+02, -3.99080582e+01, 1.61369843e+02, -9.59684944e+00, 1.02358696e+02, -1.57582321e+02, -1.25755014e+01, 1.63207367e+02, 2.14203663e+01, 5.07556076e+01, 5.89713058e+01, 6.77905579e+01, -2.05718277e+02, -4.78393517e+01, 1.27282997e+02, -1.83539825e+02, -1.52913637e+01, -4.58907547e+01, 1.76505775e+01, -7.79049683e+01, 1.45090012e+02, 1.55098892e+02, -9.25973892e+01, 1.77761017e+02, -9.83678131e+01, -6.64481964e+01, 2.94987831e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.57040466e+02, 1.72607136e+01, 1.13451317e+02, 1.31724152e+02, 5.70634308e+01, 4.01803665e+01, 8.03025589e+01, 9.11502380e+01, 8.60213394e+01, 4.74343147e+01, 1.81701263e+02, 3.89828720e+01, 6.68674612e+00, 1.38985901e+02, -4.06058083e+01, -4.60455818e+01, 3.11220016e+01, 8.89891052e+01, -1.20072266e+02, 2.48753471e+01, 8.90402832e+01, 1.57342285e+02, -1.36748764e+02, 1.91356964e+02, -2.03876892e+02, 1.64556408e+01, 1.62492172e+02, -5.40531807e+01, 9.26214600e+01, 1.82043015e+02, -2.46380596e+01, -1.66371063e+02, 8.09349442e+00, 1.07226311e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.13999512e+02, -6.72230453e+01, -3.45983353e+01, -3.64698982e+01, -4.86672287e+01, -9.07439880e+01, -1.20442116e+02, 3.72855034e+01, -1.97541065e+01, -3.59045982e+01, 5.63026657e+01, -1.60295486e+01, -5.74133644e+01, -1.70477570e+02, -1.10016975e+02, -3.13975849e+01, -6.54343262e+01, -9.44733658e+01, -1.21151772e+02, 1.88663071e+02, -8.48526993e+01, 1.17946342e+02, 1.31531876e+02, -6.74089661e+01, -1.15214294e+02, 9.59166870e+01, -7.53335571e+01, 1.92603134e+02, -8.94525909e+01, -3.58438950e+01, 1.44826477e+02, 1.43249054e+02, 9.11152725e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.90791290e+02, 2.28250160e+01, -6.85399780e+01, -1.41363159e+02, 3.56333237e+01, 1.42386408e+01, -6.77601395e+01, -1.86057796e+01, 6.15708542e+01, -1.19585075e+02, 8.69189148e+01, -1.09838249e+02, 8.84923172e+00, 1.74464798e+02, -5.67544250e+01, -1.20326385e+02, -8.05806808e+01, 2.22921967e+02, -2.17968979e+02, 1.89004807e+02, 9.57349167e+01, -2.05183868e+02, 1.90596207e+02, 5.41087646e+01, 1.63023239e+02, 5.73757095e+01, 1.37849838e+02, 2.21648422e+02, 6.86415558e+01, -2.21275177e+02, -1.45771332e+02, -1.14140205e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.35757111e+02, 6.38650208e+01, 1.06952454e+02, 5.43145409e+01, -8.50525188e+00, 1.34920471e+02, 7.25507812e+01, 1.18865303e+02, 5.20154991e+01, 1.99378796e+01, 9.19509354e+01, -4.35439911e+01, -8.64634399e+01, -3.92595558e+01, 1.23369957e+02, -6.94292974e+00, 3.89486542e+01, -1.44362534e+02, 1.81254868e+02, -1.75465363e+02, -1.69034271e+02, -7.36310043e+01, -6.58421249e+01, -1.87708405e+02, 7.67598724e+01, 4.31951675e+01, 2.31413555e+01, 6.13193245e+01, -1.89658859e+02, -1.47181900e+02, 1.34909073e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.47917297e+02, 1.78616638e+01, 1.58934296e+02, 7.13333130e+01, -5.60893593e+01, -2.23050461e+02, 2.60379601e+01, -2.47736053e+01, -2.75177765e+01, 6.83549347e+01, -2.72277393e+01, 2.66544933e+01, -8.99180679e+01, -1.40347159e+00, -1.28443207e+02, 8.16837845e+01, -3.64936981e+01, -1.47871841e+02, -2.10942337e+02, 1.08039078e+02, -3.79306488e+01, 8.77188873e+01, -1.98735905e+01, 1.96085281e+02, 2.17027542e+02, 1.42122635e+02, -2.43706684e+01, -4.80957603e+00, -1.82623138e+02, 7.21887436e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.03155090e+02, -2.53447247e+01, -2.34514465e+01, -1.09163857e+01, -1.62966354e+02, -1.82388287e+01, 1.47506361e+01, -8.55021439e+01, -1.55328156e+02, -8.25079498e+01, 1.39548874e+01, 2.38828640e+01, 3.02502556e+01, -3.90159035e+01, -8.52876759e+00, 3.92342257e+00, -7.94901123e+01, -2.71547928e+01, 9.41200104e+01, -1.45922150e+02, -1.09808502e+02, 1.91167648e+02, 1.72996948e+02, -3.20787849e+01, 5.90910339e+01, -1.81819260e+02, 5.75436058e+01, -1.77299164e+02, 3.62103043e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.84358826e+02, 1.70495544e+02, -5.35523376e+01, 5.06886711e+01, 3.90136604e+01, 1.41055801e+02, -7.73796310e+01, 6.21654358e+01, -4.27526512e+01, -3.05631599e+01, -6.56020355e+01, 5.04253807e+01, 4.90062408e+01, -2.14531975e+01, -2.86724949e+01, -2.50620899e+01, -5.30723572e+01, -9.16808701e+01, 1.06955696e+02, -2.73641071e+01, -7.40717545e+01, -1.12822281e+02, 5.26817894e+01, 5.83370895e+01, 9.01270905e+01, 2.22108290e-01, -2.83397636e+01, 5.08039703e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.63985748e+02, -1.66299454e+02, -1.59756839e+00, 2.89144382e+01, -3.05285587e+01, 7.30808640e+00, 2.26330414e+01, 9.40441055e+01, 1.13650970e+01, 5.14280396e+01, 4.74033318e+01, 1.16548950e+02, -3.07277126e+01, 5.21065826e+01, 5.19328690e+01, 9.85250168e+01, -1.62277267e+02, 7.81000316e-01, 2.13954642e-01, 7.35785904e+01, -9.97810211e+01, -3.87095795e+01, 1.46847290e+02, -5.50095673e+01, 6.79908142e+01, 8.49063339e+01, -5.63280449e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.47097107e+02, -1.10188896e+02, 8.62182081e-01, -6.50629654e+01, 3.07754097e+01, -9.98698120e+01, 6.63195724e+01, -6.31301498e+01, 1.11453224e+02, -7.27275372e+00, -7.53243484e+01, -1.08340279e+02, -1.58611130e+02, 1.90371456e+01, 1.11101616e+02, 8.90488129e+01, -6.59333038e+01, -2.83477859e+01, -1.43370895e+01, -1.65082214e+02, -6.06839333e+01, -1.14165550e+02, -1.50885178e+02, -8.66784515e+01, 1.54167236e+02, 1.20180946e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.15394073e+02, 3.06696129e+01, 5.50704384e+01, -4.82849503e+01, -1.14946533e+02, 8.67668533e+00, -2.95490971e+01, -1.34315491e+00, -6.27485771e+01, -1.31513763e+02, 9.31466961e+00, 8.06703644e+01, -9.34525375e+01, -3.38509445e+01, 8.48860073e+00, 2.68337994e+01, 9.14846191e+01, 1.45744991e+01, -9.16205883e+00, -2.68647423e+01, -1.77040161e+02, 9.18296585e+01, -2.00833557e+02, -1.02567215e+02, 5.19820023e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.60087860e+02, 7.52494812e+01, -4.64953499e+01, -3.87508926e+01, -1.92500877e+01, 1.42845551e+02, 1.35527573e+02, -7.43256760e+01, 7.83015976e+01, -9.32493057e+01, 3.70763254e+00, -1.07885811e+02, 1.69017761e+02, -1.48661819e+02, 3.44077873e+01, -6.89747162e+01, -1.51240997e+02, 7.95042801e+01, -2.16115494e+01, -7.53912582e+01, 7.40340881e+01, -2.96504269e+01, 8.05434418e+01, 1.26134117e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.26772247e+02, 1.54839386e+02, 4.48459778e+01, 1.00559216e+01, 3.31832619e+01, 3.77432327e+01, 1.72391453e+01, -2.61789341e+01, 2.64648075e+01, -8.00206985e+01, -2.93103981e+01, 9.78795471e+01, 1.12918348e+01, 1.07477257e+02, 6.61762238e+01, 6.26224327e+01, 2.30382633e+01, 3.30317650e+01, -1.79573441e+01, -1.52185287e+02, 1.34032532e+02, 7.24128418e+01, -1.40173203e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.14616913e+02, 1.50369904e+02, -8.78017807e+01, 5.16123199e+00, 5.28900766e+00, 4.51502190e+01, 1.08885475e+02, 3.92570229e+01, 2.60932827e+01, 6.02160378e+01, -1.61176186e+01, -1.01393845e+02, 2.91625957e+01, 9.47761917e+00, -4.84323502e+01, 4.36178360e+01, -1.50135315e+02, -1.41826355e+02, -1.45892929e+02, -3.98986130e+01, -6.52234650e+01, 6.28136826e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.65258209e+02, 7.65800858e+01, -4.82785110e+01, 8.79775162e+01, -9.19598541e+01, -8.98209305e+01, 2.28172874e+01, 9.76428223e+01, -6.78336487e+01, 7.98673868e+00, 1.96662273e+01, -7.96359711e+01, -6.11294403e+01, 6.69506531e+01, -8.91816559e+01, 6.34887428e+01, -2.15656185e+01, 9.43422852e+01, -9.03825989e+01, -4.12097778e+01, 8.29623871e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.64523285e+02, -1.42162399e+02, -4.48305397e+01, 4.74396286e+01, 5.20064888e+01, 1.26392563e+02, -4.37861328e+01, 1.89431286e+01, 1.26615349e+02, 1.66416107e+02, -9.19050598e+00, 1.00718918e+01, 1.03910141e+02, -8.35421295e+01, 1.03536537e+02, 6.47289810e+01, -9.85290451e+01, -1.09196442e+02, 6.03960800e+01, 3.69674263e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.28475800e+02, -1.51773148e+01, -1.43414902e+02, -3.78321762e+01, -3.20548439e+01, 1.40971786e+02, 3.54091358e+00, -1.19204330e+02, -2.91417770e+01, 1.09832664e+02, 7.30279617e+01, -4.51084366e+01, 2.97122421e+01, -1.33098663e+02, 1.84086437e+01, 8.40926514e+01, -6.99284058e+01, -2.89016266e+01, 1.07536278e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.96404724e+02, -4.78680191e+01, 1.04329430e+02, -4.22901726e+01, -7.22087708e+01, 1.03721489e+02, 1.09479889e+02, 1.41142931e+01, -8.35204849e+01, -2.97870846e+01, 5.09345245e+01, 2.60250072e+01, -1.14841820e+02, 6.60924988e+01, 7.53749008e+01, -2.77363110e+01, -1.17998138e+02, -6.34644508e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.53902878e+02, 1.80795021e+01, -2.77735710e+01, -1.15316833e+02, -6.32301445e+01, -1.52823305e+01, 1.16486273e+01, -1.02094856e+02, -5.56177807e+00, -9.61469040e+01, -8.42064209e+01, -5.46970711e+01, 1.18901016e+02, -3.55979657e+00, -1.64822121e+01, -1.00847168e+02, 5.13725548e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.21081299e+02, -4.93109894e+01, 6.40557022e+01, 8.39579010e+01, -1.01651054e+02, 2.79850235e+01, 3.07547340e+01, -7.15649033e+01, 1.02653526e+02, -2.74676781e+01, 3.52712517e+01, 3.68679428e+01, 9.29326401e+01, 9.71526566e+01, -9.62595978e+01, 4.47455063e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.07046326e+02, 1.02328506e+02, 8.73524094e+01, 3.55537148e+01, 5.59487877e+01, 1.58228855e+01, -5.40801477e+00, 3.61392937e+01, -5.46436787e+00, 9.07622910e+01, -1.69763489e+01, -7.46232452e+01, 8.40510845e+00, 1.38300695e+01, -2.94373589e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.19533447e+02, 5.41188316e+01, 2.40007095e+01, 6.17631302e+01, 6.16453438e+01, 8.43252029e+01, -4.19831429e+01, -7.22435074e+01, -5.32018738e+01, 9.95970764e+01, 7.29008713e+01, -5.07393265e+00, -9.60234756e+01, -2.69025803e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.57077820e+02, 3.95085716e+01, 8.39262924e+01, -8.65759048e+01, -5.70128822e+01, 1.14390396e+02, -5.24976120e+01, 7.73753433e+01, 8.98064651e+01, -4.85661430e+01, -2.34977455e+01, 4.29638901e+01, -6.79163895e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.42804245e+02, -7.54160995e+01, -1.50951176e+01, 2.21350479e+01, -7.66618271e+01, 1.63377895e+01, 8.30147247e+01, -1.21237518e+02, -4.08209343e+01, 9.51064377e+01, -9.43172836e+00, 1.06132797e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.38524384e+02, -5.68965950e+01, 1.15567307e+02, -1.03746132e+02, 1.04500298e+02, -4.64096375e+01, 1.48468599e+01, 5.34210434e+01, 7.93538513e+01, -5.13725052e+01, 1.26648178e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.25870895e+02, 6.13129997e+01, 6.95988941e+00, 7.68052139e+01, 4.19798546e+01, -2.61981602e+01, -9.20763168e+01, -1.01749754e+01, -1.31878910e+01, -1.07527817e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.96463501e+02, -4.74901733e+01, 1.54591246e+01, -9.32263641e+01, 6.81210852e+00, -3.62514000e+01, 3.61885109e+01, 9.16588974e+01, -8.35564651e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.03526001e+02, -8.85839310e+01, -4.00509977e+00, 8.72620468e+01, 8.41325836e+01, -9.03176956e+01, 8.82738781e+00, 2.48250237e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.62849503e+02, 4.60798416e+01, 7.20232010e+01, -7.09582186e+00, 3.91912613e+01, 4.86588631e+01, 5.88574219e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.58475891e+02, 6.12995567e+01, 2.97840347e+01, 2.45732059e+01, 1.01475506e+01, 1.20885143e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.42337921e+02, -3.24553909e+01, -5.28998985e+01, 5.29429665e+01, -4.11776276e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.34365982e+02, 1.78312855e+01, -1.49523859e+01, 3.81384315e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.64757233e+02, 8.10821533e+01, 1.83620644e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.35277496e+02, 6.48849106e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.87721817e+02 }} };
-
-const std::array<float, 50> test_solution_normal = { { 0, 0, -0, 0, -0, -0, -0, 0, -0, -0, -0, -0, -0, 0, -0, 0, 0, -0, 0, -0, -0, -0, 0, -0, -0, -0, -0, -0, -0, -0, -0, 0, 0, -1, 0, 0, -0, -0, -0, -0, 0, -0, -0, -0, -0, 0, -0, -0, 0, -0 } };
-
-const std::array<std::array<float, 50>, 50> test_mu_knapsack_big = {{{-1.55418789e+02, -4.34760819e+01, -1.86721311e+01, -3.69582084e+01, 7.77769539e+01, 2.33047756e+01, -5.23553172e+01, 3.15277196e-01, -4.20155120e+01, -7.39743249e+01, 7.48815511e+01, -1.84662358e+00, 2.99835047e+00, 5.79723987e+01, -5.08818789e+01, -7.54670659e+01, -7.31829148e+01, 6.60087500e+01, -9.67707964e+00, 6.35573090e+01, 3.52080983e+01, -7.58981592e+01, -5.48196267e+01, 5.74061866e+01, -3.35866727e+01, 3.48477815e+01, -2.67020482e+00, 5.12614983e+01, -3.88434374e+01, -1.84018935e+01, -2.96424906e+01, -2.52800837e+01, 6.15626982e+01, 6.71347400e+01, 5.50834301e+01, -5.21429877e+01, 1.83118143e+01, 1.23344160e+01, -6.60730923e+01, 2.16190077e+01, -5.40475193e+01, 3.21003659e+01, 3.20746290e+01, 1.96115284e+01, -5.77472007e+01, -4.86363331e+01, 5.07467601e+01, -2.21273118e+01, -3.52402694e+01, 4.58117068e+00},
- { 0.00000000e+00, 1.53674430e+02, -6.42773891e+01, -7.10710176e+01, 3.09839264e+01, -7.40838909e+01, 6.91708827e+01, 3.23220342e+00, 4.46029322e+01, -3.80487100e+01, 2.10285892e+01, -4.56958518e+01, -6.52590252e+01, -2.38529420e+01, -6.18720028e+01, 2.53652327e+01, -2.53634024e+01, 1.36899927e+01, 2.19768409e+01, -1.94487609e+01, 6.95152095e+01, 6.39289532e+01, -6.12811292e+01, -1.19486628e+01, 5.97938323e+01, -6.46792704e+00, 7.80285678e-01, -1.90490435e+01, -1.61820055e+01, 3.50218171e+01, 1.28405269e+01, 2.47856525e+01, -4.22548833e+01, 3.75843624e+01, 1.81605471e+01, 4.36378857e+01, 6.04207607e+01, 2.49392226e+00, -1.22075724e+00, 9.85791686e+00, 6.65209082e-01, -1.39476806e+01, 2.01561132e+01, -4.83773885e+01, -4.71296496e+01, 4.68814676e+01, 5.78773601e+01, 7.42413114e+01, 7.01412177e+00, -2.64899545e+01},
- { 0.00000000e+00, 0.00000000e+00, -1.41653693e+02, 5.54334150e+01, 5.71335095e+01, 2.12826732e+01, -5.21874032e+01, -5.01677350e+00, -4.55013966e+01, 5.68237939e+00, 3.82703686e+01, -2.54303963e+01, 5.06424789e+01, 5.92270940e+01, -6.58998427e+01, -5.62165363e+01, -6.77372084e+01, -6.21126130e+01, 5.88553504e+01, -3.57959946e+01, 1.24482007e+01, -5.67087638e+01, -3.75451543e+01, -6.68451827e+01, 3.64462318e+01, 4.15453984e+01, 6.59851267e+00, -4.98380132e+01, 3.06552319e+01, -2.22268007e+01, 3.46064604e+01, 1.16896070e+01, 8.64974326e-01, 9.29214387e-01, -5.25142509e+01, -4.78795047e+01, 1.35075240e+01, -4.86299544e+01, 4.41089278e+01, -1.69649020e+00, 6.02414542e+01, 5.08009223e+01, -4.03059312e+01, 3.85615304e+01, 3.67207802e+01, -4.74203969e+01, -6.99999804e+01, -5.50276764e+01, -3.69551855e+01, 3.49314217e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.43851096e+02, -5.66728426e+01, 7.34690379e-01, 5.67391418e+01, -5.12506694e+01, 3.64094860e+01, 6.26141407e+01, 1.64368593e+00, -1.81770956e+01, -5.23423387e+01, -5.85406944e+01, -4.63273777e+01, 1.16239647e+01, -3.75846463e+01, 4.41678171e+00, -3.23439723e+01, 2.60685169e+01, 2.94700665e+01, -7.78556541e-01, -1.91719687e+01, 7.73798108e+00, 4.70558378e+01, 6.99177086e+00, -3.94921678e+01, -1.98148612e+01, -1.37656478e+00, -3.25130831e+01, -6.30492360e+01, 5.52777151e+01, -5.99661820e+01, 4.54440120e+01, 3.05349240e+01, -3.31759864e+01, 1.12118116e+01, 3.79571112e+01, -5.10401946e+01, 6.44509917e+01, 4.48978987e+01, 7.18518854e+01, -8.67029691e+00, 6.77438051e+01, -2.10402048e+01, -3.29999323e+01, 6.46136879e+01, 6.87656091e+01, -3.65352495e+01, 2.79505981e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.34861754e+02, 6.55299310e+01, -3.49993186e+01, -5.88857616e+01, -3.22919433e+01, -5.99448168e+01, -4.68704023e+01, -3.09001007e+01, 1.26425273e+01, -5.20430863e+01, 2.09246853e+01, 3.02139393e+01, -4.46445156e+01, -4.51565675e+01, -3.79876965e+01, 5.60734747e+01, 2.22768104e+01, -6.58763887e+01, -3.27074942e+01, 4.47274960e+01, 2.63152182e+00, -9.62953952e+00, 3.31793513e+01, 4.64571514e+01, -2.32612389e+01, 6.95218525e+00, 4.29381577e+01, -1.85384512e+01, -2.80301118e+01, 1.06549587e+00, -6.02981771e+01, 5.15728395e+01, 5.55949802e+01, 2.66203527e+01, 1.66904477e+01, 5.42296220e+00, -5.13097766e+01, -3.72018062e+01, -6.64922870e+01, -1.14982557e+01, 1.75537012e+00, -2.44124633e+01, 1.56308984e+01, 3.08480757e+01, -5.62337915e+00, 5.72894613e-01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.35306322e+02, -4.38850026e+01, -1.75135175e+01, -9.85725331e+00, 3.20044177e+01, -4.06448648e+01, -1.56964472e+01, -4.14048225e+01, 2.36480138e+01, -1.11344393e+01, 1.46473493e+01, -5.12889129e+01, 3.56352385e+01, -1.62287554e+01, -6.55020205e+01, -1.63180834e+01, 2.76576615e+01, 2.41310318e+01, -4.38935315e+01, -5.04869133e+01, -5.11476455e+01, 3.33715010e+01, 4.49575817e+01, -4.77782176e+01, 6.21461061e+01, 1.57653901e+01, -4.96518590e+01, 4.58984310e+01, -6.75441431e+01, -9.99035122e+00, 6.19251058e+01, 4.03974969e+00, 2.97766390e+01, 1.28404873e+01, 5.51627935e+01, 5.45509431e+01, -6.46054572e+01, 3.51733320e+01, -5.49871355e+00, 9.95629844e+00, -9.33552137e+00, -1.70859560e+01, -5.04418442e+01, 1.39073094e+01, 6.75374735e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.31979577e+02, -5.36560562e+01, -6.22963027e+00, 5.75412955e+01, 5.13997630e+01, -2.69963950e+01, 6.27515198e+00, 7.41682901e+00, -3.31873093e+01, 6.00797479e+01, 6.15780431e+01, 2.85897509e+01, 5.55398856e+01, 2.97744765e+01, -1.81392475e+01, -3.68911243e+01, -3.33200175e+01, -1.88236708e+01, -2.28431068e+01, 3.48891585e+01, -7.92469907e+00, -5.22903242e+01, 2.95781404e+01, 3.83758028e+00, -2.01607931e+00, 5.97656313e+01, -3.76290752e+01, -6.29470659e+01, 6.13034236e+01, 7.79165349e-01, 6.50874065e+01, -5.07413159e+01, 4.99147644e+01, -1.34249624e+01, 1.57486310e+00, 1.12631335e+01, -3.80442075e+01, 5.62892626e+01, -4.61044005e+01, -1.80760668e+01, -1.25819027e+01, -4.30767518e+01, -4.08191721e+01, 5.81090099e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.29608741e+02, 3.15920601e+01, -1.91567338e+01, 2.89970308e+01, -2.23836568e+01, -3.96307202e+01, 6.29415863e+01, 3.87198130e+01, -1.12249957e+01, 1.92822532e+01, -3.40156919e+01, -3.92529488e+01, 5.76071316e+01, -2.56212832e+01, 3.05560380e+01, 4.26420507e+01, 3.41348134e+01, 2.24530098e+01, -6.02110041e+01, 4.03949711e+01, -6.10313423e+01, -4.41275517e+01, 4.67969189e+00, 3.05880733e+01, 4.66166049e+01, 6.00002727e+01, -4.12269033e+01, 4.59070293e+01, 4.96771950e+00, -1.08105986e+01, 1.02856736e+01, -5.83229079e+01, 4.84838746e+01, -4.62902384e+01, -4.19376422e+01, -4.64662516e+01, -4.39818082e+01, 7.85911457e+00, -6.37588379e+01, 1.63078747e+01, 5.08067197e+01, -4.77268793e+01, 3.60036799e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.25209608e+02, -4.03102064e+01, 4.21132288e+00, -6.06319480e+01, 3.37833165e+01, 3.35175154e+01, 2.23223596e+01, 3.35865589e+01, -3.24334609e+01, -1.37414043e+01, 3.87812411e+01, 4.59105100e+01, -6.09063723e+01, 4.93979200e+01, -5.08881354e+01, 3.92571700e+01, 3.80193844e+01, -2.96530531e+01, 2.45199401e+00, -3.36945184e+01, -3.28165106e+01, -2.88504102e+01, -1.51595660e+01, 3.14662639e+01, -4.28826008e+01, -3.54839241e+01, -4.90943917e+01, 4.96458997e+01, 2.74825363e+01, -4.69881222e+01, 3.27650668e+00, -4.20523250e+01, -5.06902099e+01, -4.67468453e+01, -4.63688511e+01, 4.43218214e+01, -5.98123372e+01, -5.69093890e+01, 2.80828719e+01, 6.95969177e-01, 1.62429320e+01, 2.94924660e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.32312952e+02, 5.95417566e+01, 6.42192939e+01, 1.39751387e+01, 5.18744011e+01, -5.01915856e+01, 2.26185131e+01, -4.30306097e+01, 5.07039796e+01, 2.56074890e+01, -6.18487469e+01, 2.15028202e+01, 4.90871795e+01, 9.01065878e+00, -3.95106322e+01, 6.55209292e+01, 6.49123167e+01, -4.95341031e+01, 4.95369853e+01, -2.33344882e+01, 5.51117098e+01, 8.47870909e+00, -1.48609762e+00, -1.67744045e+00, -5.38954510e+01, -2.34243906e+01, 6.79681486e+00, -6.27498782e+01, 2.05087193e+01, -1.46790404e+01, -6.38576870e+01, 5.56387341e+01, -3.56128668e+01, -8.54635429e+00, 5.00098424e+01, -6.16017883e+01, -3.65702659e+01, -5.27036877e+01, -2.17099190e+01, -2.26257886e+01, 1.89219597e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.32040410e+02, 6.43416336e+01, -1.83173787e+01, -5.87556413e+00, -5.86254799e+00, -3.31164773e+01, 9.18010138e+00, 4.95856529e+01, 1.95275615e+01, -6.26327605e+01, 5.91692265e+01, 5.43359155e+01, 5.07363835e+01, 2.74737780e+01, -5.60548830e+01, 4.26353795e+01, -4.30053987e+01, -4.51642048e+01, 3.54701800e+00, -3.06530885e+01, -5.41642277e+01, -2.98080709e+01, -6.59542929e+01, 5.97357163e+00, 1.79187344e+01, -3.81709802e+01, 1.01688168e+01, -2.27521374e+01, 2.22420928e+01, -1.78773465e+01, 1.03286898e+01, -5.78248934e+01, -5.08571756e+01, -1.06330360e+01, -5.19092226e+01, 5.08364692e+01, 4.61643053e+01, -4.41751442e+01, 9.83114206e-01, -6.46778192e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.14621737e+02, 5.71583182e+01, 4.34104110e+01, 4.58615505e+01, 5.71073634e+01, -1.78913724e+01, 4.92239088e+01, -1.58244225e+01, 3.52691239e+01, -2.65920813e+01, -2.39142630e+00, 4.80563207e+01, -7.76149463e+00, 2.42204438e+00, -4.89555734e+01, -8.03281234e+00, 3.62251040e+01, -3.16686680e+00, 4.24826534e+01, -2.49681136e+01, -2.84752643e+01, -3.22849443e+01, 5.95018337e+00, -3.77882296e+01, 4.11997886e+01, -1.77831203e+01, 5.19339525e+01, -5.65749642e+01, -4.03803614e+01, 3.09563648e+01, -4.83663219e+01, 4.99585727e+01, 2.43323464e+01, -9.27659215e+00, 5.02629397e+01, -1.95253086e+01, -5.36958471e+01, -1.58267019e+00, -4.93186617e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.07589532e+02, 5.20608659e+01, -4.32202617e+01, 2.39144702e+01, 2.37961541e+01, 5.71198276e+00, -2.26880284e+01, -3.26315881e+01, -3.46047167e+01, 3.21569677e+01, -9.71743125e-01, -2.65192596e+01, 3.30303920e+01, 3.37200463e+01, -7.68973953e+00, -1.47565606e+01, -3.24575364e+01, -2.87932790e+01, 5.32154896e+01, 9.67156203e-01, -4.73977210e+01, 4.67395007e+01, -2.27631874e+01, -6.14387210e+00, 5.01383546e+01, 4.58200246e+01, -2.29552104e+01, 1.91211770e+01, -2.91888367e+01, -1.99526548e+01, -2.92555987e+01, 8.22760670e+00, 2.25267862e+01, -3.33962758e+01, 4.45813781e+01, -2.18099011e+01, -4.38993246e+01, -4.24778427e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.00201374e+02, -2.08727331e+01, -3.10741935e+01, 4.45638102e+01, 4.70035527e+01, -3.52584557e+01, 2.76496624e+01, -2.77645142e+01, 4.98576405e+01, 3.16095529e+00, -4.71383868e+01, 2.06208533e+01, 3.83967044e+01, 1.94834426e+01, -2.04937937e+01, 1.99247015e+01, 1.50322634e+01, 2.39429737e+01, -4.15319707e+01, 4.02264770e+01, 2.68353846e+01, -9.90923259e+00, 2.89317904e+01, 2.21544392e+01, 1.04123301e+01, 1.42908607e+01, -2.84879653e+01, -4.19619221e+01, -3.49151544e+01, -8.81622836e+00, -2.21497863e+01, -1.95262923e+01, 3.96790399e+01, -4.42901616e+01, 8.18339970e+00, -2.20293397e+01, -1.33003653e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -9.93963711e+01, 3.72140222e+01, 3.69201454e+01, -3.98654359e+01, 4.85222805e+01, 3.36205147e+01, -3.05993083e+01, 4.73944586e+01, 4.48451444e+01, -3.39502469e+01, -1.28687184e+00, -1.31982236e+01, 3.19571811e+01, 3.46749620e+01, -1.38152003e+01, 6.87500378e+00, -3.75254393e+01, -2.62795031e+01, -4.09536492e+01, -2.70769598e+00, 4.83682110e+01, -7.66121266e+00, 8.68657598e+00, 2.31323643e+01, -1.94528600e+01, -1.99592960e+00, -7.16555785e+00, 7.98894457e-01, -3.42986211e+01, -4.53578787e+01, -1.77374000e+01, -1.77611935e+01, -3.66257643e+01, -2.77914599e-01, -4.17195107e+01, -3.36688575e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 9.24062275e+01, 3.38125127e+01, 1.46670782e+01, -3.78895617e+01, 1.20396419e+01, 3.43379214e+01, 3.55982262e+01, 2.72232301e+01, -3.67368676e+01, -1.63251047e+00, 1.88688586e+01, 2.62550504e+00, -2.65220449e+01, -1.57775863e-01, -1.75231547e+00, -6.45838917e+00, -2.15884416e+01, -3.60842473e+01, 3.51717559e+01, 3.12989981e+01, -2.74143645e+01, -3.65735285e+01, 2.88838411e+01, 3.66427136e+01, -2.46910522e+01, 3.55823606e+01, -7.39139150e+00, -7.84941149e+00, 7.40571390e+00, -2.30811252e+01, 7.14388357e+00, 2.88039031e+01, 1.04072886e+01, 1.64346170e+01, 3.30137271e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 9.27262973e+01, -4.57081377e+01, 7.54714985e-01, -3.49062063e+01, 6.43372916e+00, -3.56347096e+01, 7.25172744e+00, 1.85046848e+01, -6.70898787e+00, 1.72065802e+01, 1.06248116e+01, 2.41554337e+01, -4.59701677e+01, -2.40689732e+01, 2.28511902e+00, -4.34996164e+01, 3.58530819e+00, -1.20683726e+01, -3.87580860e+01, 1.36458617e+01, 3.48969496e+01, -2.07506701e+01, 1.04179491e+00, 2.85540117e+00, 3.88747165e+00, 4.16045560e+01, 1.65263497e+01, 2.52230744e+01, -4.55331333e+01, -4.47610871e+00, 3.88074015e+01, 1.48898320e+01, 2.59625160e+01, -3.94026393e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -8.16731092e+01, -2.42904062e+01, 1.39166211e+01, -1.78506042e+01, 1.23265781e+01, 8.08275006e+00, 3.36430965e+00, 1.56014921e+01, 2.36771823e+01, -2.95436879e+01, -3.54296000e+00, -2.93689661e+01, -1.58262304e+01, -2.57675224e+01, 1.97330386e+01, -1.04023409e+01, -2.33196476e+01, 2.14044342e+01, 3.10036722e+01, -4.62868694e+00, 3.00442088e+01, -3.96308387e+01, -7.20141350e+00, -1.34933394e+01, 3.21980954e+01, -7.53300344e+00, -1.28329820e+01, 2.91142211e+01, -5.43460634e+00, 5.28449068e+00, -1.85322301e+01, 2.31603678e+01, 1.75122670e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -8.26219209e+01, 5.12769761e+00, -2.60359319e+01, -9.52541952e+00, 1.06669092e+01, -3.92168641e+01, -3.29118535e+01, -2.55028082e+01, -3.38019375e+01, 1.82725412e+01, -2.68776554e+01, -1.64550163e+00, -8.91948299e+00, -3.36125792e+01, 2.93312598e+00, -2.73517597e+00, -3.17952447e+01, -3.78303121e+01, 3.28857427e+01, -1.03732636e+01, 2.20342879e+01, -2.89849481e+01, -2.40171264e+01, 1.51730697e+00, -6.74556948e-01, -3.57514853e+00, 3.12771600e+01, -4.97835033e+00, 3.32742408e+01, -3.67595230e+01, -1.21931035e+01, 5.26910525e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -8.40214355e+01, -3.62521298e+01, 4.12021620e+01, -1.69984913e+01, 3.93964074e+01, -3.39834262e+01, -1.06636837e+01, -4.16699967e+01, 3.10236420e+01, 2.52921022e+01, 3.74684020e+01, -2.77918680e+00, 1.09975456e+01, -3.00899308e+00, 1.77943101e+01, -3.62172970e+01, -3.00638565e+01, 2.11301672e+01, -2.43970264e+01, -5.17306557e+00, -8.75765517e+00, 3.86269068e+01, -1.26406945e+00, -1.38822714e+01, 3.30381135e+00, -2.73568070e+01, -2.96438367e+01, 2.91207075e+01, 3.89024238e+01, 1.60322915e+01, 1.15308696e-01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -7.71788958e+01, 3.30009699e+01, 3.50305321e+01, 6.32412805e+00, -1.72047261e+01, 8.64616823e+00, -2.99915070e+01, -2.12533048e+01, -3.01399535e+01, 2.64390758e+01, 3.53774938e+01, 3.01519789e+01, -7.61494281e+00, -5.68063359e+00, -2.75074139e+01, -1.77282883e+01, 1.12191428e+01, -2.84858458e+01, 2.23616434e-01, 2.73695900e+01, 1.15370267e+01, 2.67446673e+01, -1.89030951e+01, 8.39852149e+00, -3.66760514e+01, 7.41335778e+00, -6.49951852e+00, 7.15166811e+00, -2.30614233e+01, -3.16601225e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 7.47894680e+01, -1.95340898e+01, -5.35672794e+00, -3.06748646e+01, -2.56422263e+01, -2.68999880e+01, -5.65016905e-01, -2.46340628e+01, 3.66339054e+01, 3.48707112e+00, 5.69827366e+00, -1.87182566e+01, -3.61597792e+01, 2.95233308e+01, -4.92695854e+00, 2.33191809e+01, -5.37719488e+00, 1.53509154e+01, 2.00991560e+01, 1.39157296e-01, -1.99520409e+01, -5.58918463e+00, -1.71277988e+01, 2.69100244e+01, 1.65827970e+01, 3.69604750e+01, 2.26004636e+01, 2.85400843e+01, -2.15387194e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -7.51976921e+01, -3.13682364e+01, -6.36090455e-01, 3.23656383e+01, 3.53476416e+01, -2.81124000e+01, 3.13745758e+01, -2.06759153e+01, 1.13153858e+01, -1.93815564e+01, 2.40880008e+01, 1.76368427e+01, 2.03227763e+01, -1.85741109e+01, 3.50298452e+01, -1.98243113e+01, 2.86801373e+00, 9.03431925e+00, 8.69874540e+00, -4.20185261e+00, -3.06338290e+01, 3.62371197e+00, 1.14209933e+01, -1.61671639e+01, 2.14154405e+01, 3.28696589e+01, 1.15805139e+01, 2.07108404e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 7.43483731e+01, -3.27079619e+01, 2.10071511e+01, 8.00126339e+00, 3.59298996e+01, -2.83076734e+01, -2.93146514e+01, -2.90836132e+01, 6.45775167e+00, -2.06409279e+01, -3.59821957e+01, -2.69978161e+01, -3.01926122e+01, -2.27804311e+01, -2.52234716e+01, -1.60203310e+01, 3.25929328e+01, -1.00034429e+01, 3.33611961e+01, 2.87881729e+01, -3.68069897e+01, 9.95521359e+00, -2.14343232e+01, -1.10586826e+01, 1.00001547e+01, -3.21096299e+01, 1.56954461e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.95598997e+01, 2.14470959e+01, -1.22112037e+01, -3.12608666e+01, 2.90859503e+01, -3.19834736e+01, 2.63510004e+01, 1.59856272e+01, 3.22416118e+01, -5.97683247e-01, -2.33000210e+00, -3.12759451e+01, -2.22253291e+01, 6.93101257e+00, -1.19554985e+01, -2.31429255e+01, 1.01490400e+01, 2.15717389e+01, -2.71164742e+01, 2.66143151e+01, 2.66672686e+01, -1.44755918e+00, 3.17446153e+01, 2.89189441e+01, -3.47277624e+01, -8.54724857e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.68332092e+01, 2.58168766e+01, -1.00361517e+01, 2.98107937e+01, -2.66997626e+01, -2.27692036e+01, -2.33147326e+01, 2.22287764e+01, 2.20509292e+01, 6.15347418e+00, 1.82841085e+01, 1.30422777e+01, -2.03614866e+01, 1.67206197e+01, 3.19823686e+01, 1.20510602e+01, -2.30533267e+01, -1.83822973e+01, 3.64513612e+00, -2.25354363e+01, -3.14552951e+01, 1.13928789e+01, 2.77257014e+01, 3.03765045e+01, -2.78089042e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.60424938e+01, 1.62049018e+01, -2.59657845e+01, -2.50390446e+01, 2.14163557e+01, -1.15191726e+01, 2.43261406e+01, -2.51950726e-01, -2.28611232e+01, 1.95797696e+01, 4.77109895e+00, 3.08397424e+01, 1.05158035e+00, -1.38193487e+01, 3.10723471e+01, 7.81314690e+00, 4.95678227e+00, 7.80605133e+00, -2.61903422e+01, -2.36256375e+01, -2.30063300e+00, -3.00082180e+01, 2.38221884e+01, -3.02972546e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.47172349e+01, -2.89192293e+01, -1.41567814e+01, 2.56429931e+01, 2.64102857e+01, -2.77910382e+01, -2.36045934e+01, -1.44257935e+01, -5.22973532e-01, -2.39714388e+01, 3.14538848e+01, 1.72815913e+01, 1.28418388e+01, 2.09871486e+01, -4.43321806e-01, -1.31159515e+00, -2.40460862e+01, 8.73870306e+00, -2.52522653e+01, -7.41038373e+00, 2.63837047e+01, 1.72293656e+01, 2.70562510e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.84746613e+01, -2.47405055e+01, -5.47216848e+00, 2.83301761e+00, -8.42969519e+00, 5.28566216e+00, 1.04663334e+01, 1.72153057e+01, 1.68714824e+01, 1.88370092e+00, -5.69494702e+00, 1.09311757e+01, -6.75753357e+00, 1.78412854e+00, 4.59113789e+00, -1.46440300e+00, -4.05547245e+00, 1.32730796e+01, 4.69430327e-01, -1.43079381e+01, -5.59910100e+00, 1.18923512e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.97131168e+01, 2.33763537e+01, -1.33145194e+00, 2.41054494e+01, 1.65525243e+01, -4.31561856e+00, 8.21238839e+00, -1.65187535e+01, -1.12787205e+00, -1.69713869e+01, -1.48809517e+01, 7.24650440e+00, -6.91839594e+00, 1.57832731e+00, -1.45375433e+01, -1.22232694e+01, -2.42522941e+01, -2.87821537e+01, 1.76014003e+00, -2.11961991e+01, 1.36007523e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.97685926e+01, -2.60515567e+01, -2.96711898e+01, 1.17968506e+01, 2.39353511e+01, 1.97440563e+01, 1.55998902e+01, 2.55572114e+00, -1.93063184e+01, 1.65086760e+01, 6.47133121e+00, -5.05057999e+00, 1.93062575e+01, 2.92415119e+00, 1.97515360e+01, 2.07552921e+01, -3.42101219e+00, 1.31026782e+01, -1.77644659e+01, 1.99694988e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.48541057e+01, -2.66934300e+01, 2.22001752e+01, -2.62124431e+01, 4.73434761e+00, -7.32526779e+00, -1.76523075e+01, -2.08937330e+01, 2.16250025e+01, -4.40697368e+00, 2.67434082e+01, -1.55319747e+01, 2.35827561e+01, 2.71467698e+01, 8.60171818e-02, -4.29522644e+00, 2.05872503e+01, -1.94089029e+01, 1.82836179e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.98033603e+01, -2.46291452e+01, -2.51829880e-02, 1.86530176e+01, 2.41120076e+01, -2.12092029e+01, -8.00000272e+00, 3.37795979e+00, 2.26005899e+01, -1.34099211e+01, -5.60574359e+00, -2.04190241e+01, -1.40676173e+01, 3.38769259e+00, -1.95448255e+01, -2.42795907e+01, 1.34944729e+01, 8.50336639e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.49827111e+01, 1.67216897e+01, -2.12513246e+01, 2.24368129e+01, -4.73868443e+00, -1.86074331e+01, 2.16754130e+01, -6.12549324e-01, 1.04171051e+01, 6.95436608e+00, -9.16721941e+00, -1.15567549e+01, -2.04626213e+00, 1.23365240e+01, -4.69516047e+00, 1.46402455e+00, -5.41485846e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.16920001e+01, -1.95274645e+01, 1.74981292e+01, 1.86938890e+01, -5.54600820e+00, -1.36975705e+01, -1.53149083e+01, 1.65396631e+01, 1.44703546e+01, -3.70700517e+00, -1.84229033e+01, -4.33333945e+00, -5.61760568e+00, -1.23742864e+01, 7.43297751e+00, 2.02636465e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.85093194e+01, -1.68739583e+01, 1.57498315e+01, -9.89521943e+00, 8.11179022e-01, 4.91582051e+00, -1.18123065e+01, 4.89352152e-01, 1.84546751e+01, -9.13624691e+00, 5.63286969e+00, -1.89998721e+01, 1.51016538e+01, -1.48724993e+01, -5.42440273e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.29846494e+01, -2.03056710e+01, -1.79100607e+01, 3.19866338e+00, -2.02161804e+01, -1.78719397e+01, -1.20409676e+01, -1.16047746e+00, 7.02550238e+00, -2.51106032e+00, 1.54560774e+01, -5.86414956e+00, 5.31087452e+00, -5.59769097e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.89217814e+01, -9.14519734e+00, -1.80565099e+01, -1.87942505e+01, -2.29069300e+00, 6.43570831e+00, 1.21392393e+01, -1.57231093e+01, -5.07475659e+00, -8.37708124e+00, 2.90506746e+00, -1.34679415e+01, 1.71959484e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.92030247e+01, 9.96939826e-01, 1.58712225e+01, 2.11371255e+00, -1.57933187e+01, -1.16301598e+01, -1.92226251e+01, -1.34876777e+01, 1.55028575e+01, -8.46522126e+00, -1.84735066e+01, 1.02830046e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.93968739e+01, -1.51830899e+01, 1.78639381e+01, 1.40483422e+01, -1.55391510e+01, 1.55762783e+01, 9.52986876e+00, -1.93546550e+00, -5.97945107e+00, 3.75018271e+00, 7.73503737e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.62442471e+01, 1.72799801e+01, 1.32142559e+01, -5.74620464e+00, 2.62661386e+00, 1.15916162e+01, -1.75541101e+01, 1.36352469e+01, 1.60502291e+01, -2.51997641e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.36379150e+01, -1.52010726e+01, 2.68938612e+00, -1.53995237e+01, 7.41461917e+00, 4.15527213e+00, -6.63000271e+00, 9.80895488e+00, -9.98768306e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.32740560e+01, -1.26778877e+01, 6.32603124e+00, 1.17969463e+01, -1.59942418e+01, 1.86714530e+00, -6.42831084e+00, -7.94348164e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.06339887e+01, 1.43098970e+01, 2.06089920e+00, -1.03224755e+01, -1.38377813e+01, -1.06347458e+01, 2.89304780e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.99313298e+01, 1.11322977e+01, 5.68209605e+00, 1.28550011e+01, -1.33946331e+01, 1.11337943e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.87550570e+01, -1.40355694e+01, -1.17392345e+01, 1.37179852e+01, 3.28147346e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.73819930e+01, -1.07083939e+01, -2.50331328e+00, -4.36615067e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.84916995e+01, 1.03099884e+01, 2.60459029e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.64828936e+01, 7.40398109e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.89637400e+01}}};
-
-const std::array<std::array<float, 40>, 40> test_mu_knapsack_normal = {{{-3.46584766e+02, -6.69504326e+01, 8.15384944e+01, 6.50634483e+01, -2.40633773e+01, 3.69923934e+01, -6.52625338e+01, 4.53366724e+01, 1.13383518e+02, 3.50159649e+01, -1.63368404e+02, 1.56065717e+02, -2.01393734e+01, -1.71865604e+02, 7.07157452e+01, 1.71227953e+02, 1.81225507e+01, 6.16501419e+01, -1.24399582e+02, -5.04436482e+01, -5.45725082e+01, -1.10414547e+02, -1.29434425e+02, -4.34410323e+01, -1.09525876e+02, -1.19237786e+02, -1.02924316e+02, 1.32241819e+02, -5.76280378e+01, 1.25490224e+02, 7.92533391e+01, -1.17073813e+02, 1.40193698e+02, 1.19154112e+02, -9.73383810e+01, -6.33005318e+01, 8.05430669e+01, 1.71735765e+02, -3.15449526e+01, 6.13991211e+00},
- { 0.00000000e+00, -3.43327307e+02, -1.70138047e+02, -2.04185916e+00, 6.99056985e+00, -1.60475603e+02, -1.19433102e+02, 1.20572088e+02, 4.53481096e+00, -9.98852504e+01, 3.55625232e+01, 2.15256188e+01, -6.85837094e+01, -7.94438264e+00, 6.96989427e+01, -2.17366500e+01, -6.27544392e+01, 1.38630093e+02, 7.71118559e+01, -4.00456814e+01, 3.81666497e+01, -3.06695625e+01, 5.87040132e+01, -1.64142505e+02, -1.03106262e+02, -1.67050443e+02, -1.09921849e+02, 1.44958912e+02, 3.70411028e+01, -1.49500560e+02, -1.65588476e+02, 1.68705318e+02, 9.56695568e+01, -1.50038806e+02, 5.51801338e+01, 1.08237816e+02, -9.22134434e+01, 1.71067710e+02, -6.77023101e+01, 1.07730814e+02},
- { 0.00000000e+00, 0.00000000e+00, 3.25458936e+02, -5.14368226e+01, -1.21129096e+02, 1.07816029e+02, -8.02734325e+01, 1.07362901e+02, 1.74197784e+01, -3.34033861e+01, -1.52402170e+02, -7.06787687e+01, 5.27760461e+01, 1.24940557e+02, -1.24429257e+02, -1.75779493e+01, -4.17306202e+01, 1.49792764e+02, 2.93524440e+01, -1.44700881e+02, 5.50525963e+01, 1.23554686e+02, 3.23102953e+01, 1.53903383e+02, -1.06184588e+02, 7.03585948e+01, 1.10133863e+01, -1.03241205e+02, 1.05902775e+02, -4.20422236e+01, 8.10698350e+01, 2.10538305e+00, -1.61223151e+01, 2.02914803e+01, -1.21136434e+02, 5.43441854e+01, 8.74599567e+01, -1.23454887e+01, 1.27556458e+02, -1.27858001e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.49175932e+02, -1.16750384e+02, 6.35175716e+01, 1.41065480e+02, 2.78517550e+01, 1.46481588e+02, 7.10397618e+01, 4.68536245e+01, 5.27319068e+01, 1.29402015e+02, 8.28450776e+01, -3.73767445e+01, 3.51405036e+01, -1.46348168e+02, -6.93454327e+01, -1.10039411e+02, -1.11471500e+01, 9.44442610e+01, -1.18758439e+02, -1.73514978e+02, -1.26058426e+02, 6.51052751e+01, -3.16158840e+00, 5.78400129e+01, 8.14962315e+01, -1.99767887e+01, -9.37532996e+01, 1.21181402e+02, 2.16252228e+01, 1.46282769e+02, -4.14140621e+01, -1.13306685e+02, 8.14865742e+01, -4.51003757e+01, -8.11669985e+01, 1.63319283e+01, 1.61778395e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.36471657e+02, -7.51375715e+01, -4.64512554e+01, -1.07086107e+02, -1.39163243e+02, -1.27453069e+02, 5.14427990e+01, 8.23166124e+01, -3.51017158e+01, -3.48914906e+00, -2.39257248e+01, -1.21942691e+02, 3.31997709e+01, 6.80787309e+01, 7.16308823e+01, -1.55495485e+02, -2.75093915e+01, 5.48885006e+01, -6.48649804e+01, -1.00427717e+02, 1.66508027e+02, -1.58560546e+02, 1.62205624e+02, 3.71771046e+01, 1.53911763e+02, -1.06045953e+02, 1.61296058e+02, -1.64818523e+02, -9.64674502e+01, 1.48915628e+02, 1.37551626e+02, 8.09589210e+01, 8.55829054e+01, -1.14624794e+02, -7.45864995e+01, 1.12389892e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.31906178e+02, -1.22418602e+02, 1.82764268e+01, -1.14991157e+02, -9.93528870e+01, 6.87934197e+01, 5.57508881e+01, 3.59864598e+01, 2.90201326e+01, -5.06382175e+01, 1.59708789e+02, 9.41275253e+01, -1.26885181e+02, 5.95691009e+01, 3.20368779e+01, 1.59304575e+02, 7.27134477e+01, 1.58763349e+02, 3.40665516e+01, 1.31316222e+02, -9.86768372e+01, -4.63712322e+01, 1.73254953e+01, 1.54487661e+02, 3.49357635e+01, 1.27623225e+02, 1.43797531e+02, 1.11901856e+02, -6.41971520e+01, -5.99098561e+00, -4.83533628e+01, -1.65380821e+02, -1.59785454e+02, 3.43413630e+01, -2.98544849e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.26723445e+02, 1.40897013e+02, -1.39429199e+02, 1.06855000e+02, -9.45460435e+01, 1.47367777e+02, 7.93431315e+01, 6.22111816e+01, 6.18192101e+01, 1.21691678e+02, -1.33529192e+01, -1.42209962e+02, 1.00899492e+02, -9.60320761e+01, 6.17854204e+01, -4.67493209e+01, 9.99434150e+01, -1.19517321e+02, -7.26452615e+01, 1.28981906e+02, 2.20480121e+01, 1.39248091e+02, 1.25501387e+02, -1.09076331e+02, 2.65342731e+01, -1.51260541e+02, -5.79410513e+01, 1.05850828e+02, -3.29282993e+01, 3.78859586e+01, -1.26687170e+02, 2.49056883e+01, 6.54152380e+01, -3.13729896e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.95536824e+02, 1.21467519e+02, 5.46577592e+01, -2.01197197e+01, 1.39207936e+02, 1.28191302e+02, 1.00000839e+02, 1.41737197e+02, 6.30045157e+01, -1.44633424e+02, 1.01413850e+02, -4.70964061e+01, 6.15158398e+00, 5.67267778e+00, -1.03405536e+02, 9.11505249e+01, 1.07795422e+02, 1.32252456e+02, 1.10958359e+02, -4.68051720e+01, -4.46647471e+01, 6.29987769e+01, -1.34977310e+02, -6.60705186e+01, -8.45054634e+01, -1.35289591e+02, -9.47686733e+01, 1.40104451e+02, 1.08221933e+02, -3.58901559e+01, 3.34192492e+01, -2.88687940e+00, 1.37198101e+02},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.77405975e+02, 8.04392596e+01, -1.37879614e+02, 4.04371438e+01, -1.00881221e+02, -7.20090340e+01, -8.03006208e+01, 7.33854411e+01, -1.19188235e+02, -6.68095720e+01, -8.29900430e+01, 3.91701697e+00, 9.52424985e+01, -1.31442586e+02, 1.03091054e+01, -6.14218438e+01, -1.18013814e+02, 4.54255242e+00, -1.73113499e+01, -9.14683564e+01, 6.64500768e+01, -1.00060825e+02, -7.03118467e+01, -3.04828297e+01, -5.06594500e+01, -5.65405937e+01, -7.54913918e+01, -1.29927657e+02, 1.87774754e+01, -1.21617842e+02, -9.05558388e+00, -1.11789741e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.87593755e+02, 1.38973210e+02, 7.55081321e+01, -1.29510452e+02, -1.41373527e+02, 9.92207712e+01, -6.47789441e+01, -3.98360736e+01, -6.67176140e+01, -8.93938648e+01, -5.47491893e+01, -8.09081537e+01, -1.38827798e+02, 1.02536181e+02, 1.12146199e+02, -6.15583833e+01, 3.71294179e+01, 7.46654153e+01, 6.41987350e+01, 1.36652054e+02, 1.15977444e+02, -5.70197487e+01, 8.12270221e+01, 8.56015651e+00, -3.75658620e+01, -2.32702433e+01, 1.78494405e+01, -1.22379053e+02, 6.14239580e+01, -4.21970948e+01, 6.92049405e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.61975941e+02, -1.12547560e+02, -8.84471540e+01, 1.89070614e+01, -1.11613129e+01, 9.18634871e+01, -3.19712964e+01, 1.13508491e+02, -9.16007346e+01, -7.97003553e+00, 7.92598408e+01, 1.24192769e+02, -1.05099042e+02, -1.09762777e+00, -3.86563441e+01, -8.63338468e+01, -1.50694683e+01, -1.55729324e+01, 6.24402733e+01, -5.44992269e+01, 9.26098213e+01, -1.17290749e+02, 9.06260198e+01, -4.27090513e+01, -6.11546583e+01, -1.00324226e+01, -2.77972222e+01, -1.30022971e+02, 1.16261516e+02, 9.47548791e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.40991921e+02, 5.80315829e+01, 8.57228704e+01, -9.62882647e+01, -5.68261351e+01, 6.75287405e+01, -3.63816588e+00, 8.81549662e+01, 9.38667647e+01, 4.89219629e+01, 9.08564992e+01, 2.03790613e+01, -8.19910514e+01, 1.12670867e+02, -3.82798188e+01, 1.03958422e+02, -7.88194889e+01, 7.27946338e+01, -5.77665022e+01, 2.15066121e+01, 1.08285631e+02, 1.12639140e+02, 8.17148793e+00, -1.05669184e+02, -1.14289505e+02, -5.65099378e+01, -4.64230923e+01, 7.17576747e+01, 2.45202038e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.41454858e+02, 1.03226539e+02, -9.51862399e+01, 5.61005157e+01, 5.77533173e+01, 8.76119916e+01, 1.46066026e+00, -1.15586532e+02, 3.68431744e+01, 9.66468902e+01, -3.34604166e+01, 1.14814552e+02, 6.81291408e+01, -7.27859720e+01, -1.04088229e+02, -1.18596377e+02, 1.15826935e+02, -3.87713960e+01, 9.71171417e+01, -7.95204867e+00, -2.38907993e+01, 6.35017947e+00, 1.12321143e+02, -9.52220553e+01, 4.24712078e+01, -4.37920638e+01, -8.50788876e+01, 8.24069289e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.39769742e+02, -6.09704864e+01, 9.82981161e+01, 4.02771228e+01, -5.58106616e+00, 8.76616327e+01, 6.19236046e+01, -7.36955266e+01, -1.12979121e+02, 1.05226241e+02, -5.79551777e+01, -5.91304900e+01, 9.71142635e+01, -1.18595011e+02, 5.56256665e+01, 9.55604923e+01, -3.86507182e+01, -7.57006271e+00, 1.13461204e+02, 1.25068113e+00, 1.02210356e+02, 2.34428641e+01, -4.61086957e+00, 7.10057612e+01, -5.78995574e+01, -3.76557843e+01, 5.45575557e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.36495637e+02, -1.07351699e+02, 6.84922538e+01, 1.02797521e+02, 8.68013874e+01, -2.64712112e+01, 7.83315188e+01, -9.02987004e+01, 8.39392764e+00, 1.09096311e+02, -1.09970887e+01, -8.72240651e+01, -5.44467333e+01, -1.73838008e+01, 3.31476918e+00, -2.64964562e+01, 6.03282739e+01, 1.04532297e+02, -7.09332626e+01, -6.26371595e+01, -6.25127225e+01, 9.41553750e+01, 4.82754051e+01, 4.69310785e+01, 5.63641880e+01, 1.47187885e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.11606467e+02, -8.16098099e+01, 2.21894283e+01, 8.62710140e+01, -7.09838812e+01, -9.74894414e+01, 5.30756990e+01, 1.51863674e+01, -6.44647920e+01, 5.09548875e+01, 6.82542852e+01, -9.29507094e+01, -2.23576110e+01, -7.95144274e+01, -5.51113852e+01, -5.51401528e+01, -2.17550759e+01, -8.66539959e+01, 3.03466184e+00, -3.47847645e+00, 8.12390484e+01, -6.53163604e+01, 6.94288528e+01, -8.30663022e+01, -1.62055091e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.97640568e+02, 8.54117501e+01, 8.43088176e+01, 4.15371455e+01, -4.55653792e+01, 9.47982797e+01, -9.67487894e+01, 9.30054299e+01, 1.95559044e+01, -8.39709389e+01, -5.30272336e+01, 8.53779779e+01, 9.71342857e+01, 1.96116509e+01, -8.56941447e+01, 1.56619110e+01, 4.52742239e+01, -4.40703111e+00, -9.29905877e+01, -8.52500510e+01, 3.42314532e+01, -3.30440253e+01, 8.73830251e-01, 5.57007950e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.80793366e+02, 7.11547052e+01, 2.82139945e+01, 7.88168301e+01, 6.05527646e+01, 3.52811164e-01, -4.89885420e+01, -3.35468042e+00, 3.05294222e+01, 8.39376542e+01, 1.94980827e+01, -6.35624412e+01, -2.31613778e+01, -5.25075833e+00, -5.41591879e+01, -5.63837616e+01, -8.76249629e+00, 6.84627949e+01, 5.15866683e+01, -6.17826800e+01, 4.27138222e+01, 5.00782192e+01, 6.15757316e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.98157997e+02, -5.67101262e+00, 8.73232048e+01, -1.83975065e+01, 4.12994695e+01, 4.91582132e+01, 1.86610984e+01, -6.91653123e+01, -3.29139781e+01, -6.17671443e+01, 6.97110877e+01, -4.28987160e+01, 3.87027582e+01, 8.37446378e+01, 1.78411019e+01, -3.57018711e+01, 3.05921786e+00, 6.24120979e+00, 2.23548897e+01, -4.41664428e+01, 8.57112592e+01, 4.51362058e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.99566834e+02, 6.63227520e+01, 7.98412199e+01, -8.83852659e+01, 9.48619640e+01, 9.25623817e+01, -1.61934908e+01, 7.12787708e+01, 7.50707944e+00, 4.92774619e+01, 5.33103709e+01, 5.81892261e+01, 9.08964994e+01, -7.65291433e+01, -4.50445468e+01, 8.93814934e+01, 1.08681138e+01, 6.11946708e+01, -6.54800503e+01, -7.68324751e+00, 7.62794293e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.91642250e+02, -7.04253647e+01, 3.75258260e+01, 5.99922536e+01, 5.44965121e+01, 5.02930234e+01, -2.03219025e+00, 2.74990334e+01, 6.41271940e+01, -5.06617330e+01, 6.53121962e+01, -3.39709990e+01, 5.23000991e+01, 3.33435222e+00, 8.17595981e+01, 4.22267126e+00, 3.60053034e+01, 9.77405831e+00, -8.84306950e+01, 7.00222344e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.80070718e+02, 7.58705187e+01, -3.01114145e+01, 6.58867881e+01, -1.80012026e+01, 1.09070694e+01, 2.73744291e+01, 3.25078960e+01, 6.88972484e+01, -6.11373049e+01, 6.25995896e+01, 6.44046458e+01, 1.86576810e+01, -1.02886393e+01, 3.86105510e+01, -4.41908687e+01, -2.63844120e+01, -4.47072685e+01, 6.29310513e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.72597541e+02, 7.79327803e+01, -5.89494821e+01, -5.24850761e+01, 3.48373573e+01, 6.51709793e+01, 5.68149947e+01, -6.68346090e+01, -7.00469835e+01, -8.41604576e+01, 7.37756197e+01, 3.18853669e+00, -5.79851674e+01, 7.63743245e+01, 7.94701825e+01, -6.27945333e+01, 8.22548180e+01, 2.27610654e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.53225629e+02, -5.49609622e+01, 7.33514037e+00, 5.92709691e+01, -3.81703461e+01, -5.39033066e+01, 1.44030148e+01, -7.65038371e+01, -1.64344452e+01, -4.35780685e+01, 6.21079132e+01, 5.28218070e+00, 3.46679097e+01, -1.27101143e+01, -5.31884161e+01, 3.47616388e+01, 2.65463243e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.69357401e+02, -8.21501450e+01, -3.29345706e+01, -7.89785892e+01, -7.30412648e+01, 5.16018891e+01, -2.36891246e+01, 3.14245184e+01, -6.16050713e+01, -3.24333248e+00, 3.69326429e+00, 1.81823360e+00, -8.23771326e+01, -7.77550956e+01, -7.56134244e+01, -4.18705497e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.65514199e+02, 7.91445980e+01, -8.74613319e+00, 2.09636556e+01, -4.41058613e+01, -4.37935090e+01, -6.35121497e+01, -7.65984539e+01, 5.03633002e+01, 7.09667254e+01, 1.76280236e+01, 4.55084113e+01, 2.61339710e+00, -5.73396352e+00, 5.90940386e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.47404214e+02, -7.19384792e+01, 7.20088944e+01, 9.13255543e+00, -6.02017088e+01, -6.02434822e+01, 4.74492675e+01, -2.89797508e+00, -3.16192962e+01, 3.81268786e+01, 6.74525403e+01, -4.34105660e+01, 6.22015246e+01, -2.67179257e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.39559506e+02, -3.34839846e+01, 4.31020266e+01, -9.41331454e-02, 6.72241866e+01, 6.42028139e+01, 6.33412038e+01, -1.37558531e+01, -5.92319616e+01, 1.09836429e+01, 5.73616840e+00, 2.15422805e+01, 3.01870622e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.39357940e+02, -6.71358396e+01, 4.07270600e+01, 3.21177180e+01, 4.04600778e+01, 6.46312313e+01, 7.44357753e+00, 1.85955671e+01, 6.55061548e+01, 4.47556233e+01, 6.37835616e+01, 6.81577548e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.29914128e+02, 3.03417309e+01, -1.16899447e+01, -4.75603013e+01, 3.39538311e+01, 2.54551975e+01, 2.75281016e+01, -1.29307800e+01, 5.66870587e+01, 2.48977191e+01, 5.11675173e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.35044783e+02, 5.26872240e+01, -2.66728303e+01, -9.88520094e+00, -3.49001605e+01, 6.39937073e+01, -1.79882774e+01, -1.81251805e+01, -3.78875844e+00, 6.24030145e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.24253735e+02, -4.23652200e+01, -5.91222877e+01, -3.17831622e+01, -2.61388974e+00, 6.09208559e+01, -5.82963552e+01, -2.64480396e+00, 8.41246669e+00},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.21650014e+02, -4.92982217e+01, -2.78905307e+01, -4.73571391e+01, -5.82840223e+00, 5.89810591e+01, -4.02584026e+01, -5.20840442e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.15315768e+02, -4.16007535e+01, 3.86535176e+01, 2.92350122e+00, 2.85412150e+01, -2.09648712e+01, -4.23109955e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.10881641e+02, 5.02520644e+01, 3.13674719e+01, -4.10115618e+01, -2.25517267e+01, 2.68837975e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.13944424e+02, 5.37149691e+01, 4.06178150e+01, 1.90969104e+01, 2.28112728e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.07067540e+02, 4.11640162e+01, -8.39866573e+00, -1.15858212e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.00086288e+02, 4.89919015e+01, -3.94696367e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 9.01605401e+01, 2.04996713e+01},
- { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 9.11151257e+01}}};
-
- const std::array<std::array<float, 50>, 50> test_mu_compare = {{{1, 0.279735, 0.120141, 0.237798, -0.500435, -0.149948, 0.336866, -0.00202857, 0.270337, 0.475968, -0.481805, 0.0118816, -0.0192921, -0.373008, 0.327386, 0.485572, 0.470876, -0.424715, 0.0622645, -0.408942, -0.226537, 0.488346, 0.352722, -0.369365, 0.216104, -0.224219, 0.0171807, -0.329828, 0.249928, 0.118402, 0.190727, 0.162658, -0.396108, -0.43196, -0.354419, 0.3355, -0.117822, -0.0793625, 0.425129, -0.139102, 0.347754, -0.206541, -0.206375, -0.126185, 0.371559, 0.312937, -0.326516, 0.142372, 0.226744, -0.0294763}, 
-{0, 1, -0.41827, -0.462478, 0.201621, -0.482083, 0.450113, 0.0210328, 0.290243, -0.247593, 0.136839, -0.297355, -0.424658, -0.155217, -0.402617, 0.165058, -0.165046, 0.0890844, 0.143009, -0.126558, 0.452354, 0.416003, -0.398772, -0.0777531, 0.389094, -0.0420885, 0.00507752, -0.123957, -0.105301, 0.227896, 0.0835567, 0.161287, -0.274964, 0.244571, 0.118175, 0.283963, 0.393174, 0.0162286, -0.00794379, 0.0641481, 0.00432869, -0.0907612, 0.131161, -0.314804, -0.306685, 0.30507, 0.376623, 0.483108, 0.0456427, -0.172377}, 
-{0, 0, 1, -0.391331, -0.403332, -0.150244, 0.368415, 0.0354158, 0.321216, -0.0401146, -0.270169, 0.179525, -0.357509, -0.418112, 0.465218, 0.396859, 0.478189, 0.438482, -0.415488, 0.252701, -0.0878777, 0.400334, 0.265049, 0.471892, -0.257291, -0.293289, -0.046582, 0.35183, -0.21641, 0.156909, -0.244303, -0.0825224, -0.00610626, -0.00655976, 0.370723, 0.338004, -0.095356, 0.343302, -0.311386, 0.0119763, -0.425273, -0.358628, 0.284539, -0.272224, -0.259229, 0.334763, 0.494163, 0.388466, 0.260884, -0.246597}, 
-{0, 0, 0, 1, 0.393969, -0.0051073, -0.39443, 0.356276, -0.253105, -0.435271, -0.0114263, 0.12636, 0.363865, 0.406953, 0.322051, -0.0808055, 0.261275, -0.0307038, 0.224843, -0.181219, -0.204865, 0.00541224, 0.133276, -0.0537916, -0.327115, -0.0486042, 0.274535, 0.137746, 0.00956937, 0.226019, 0.438295, -0.38427, 0.416863, -0.31591, -0.212268, 0.230627, -0.0779404, -0.263864, 0.354813, -0.44804, -0.312114, -0.499488, 0.0602727, -0.47093, 0.146264, 0.229403, -0.449171, -0.478033, 0.25398, -0.194302}, 
-{0, 0, 0, 0, 1, -0.485904, 0.25952, 0.436638, 0.239445, 0.444491, 0.347544, 0.229124, -0.0937443, 0.3859, -0.155157, -0.224036, 0.331039, 0.334836, 0.281679, -0.415785, -0.165183, 0.488473, 0.242526, -0.331654, -0.0195127, 0.071403, -0.246025, -0.34448, 0.172482, -0.0515505, -0.318386, 0.137463, 0.207843, -0.00790065, 0.447111, -0.382413, -0.412237, -0.19739, -0.12376, -0.0402113, 0.380462, 0.275851, 0.49304, 0.0852596, -0.0130161, 0.181018, -0.115903, -0.228739, 0.0416974, -0.00424801}, 
-{0, 0, 0, 0, 0, 1, -0.324338, -0.129436, -0.0728514, 0.236533, -0.300391, -0.116007, -0.306008, 0.174774, -0.0822906, 0.108253, -0.379058, 0.263367, -0.119941, -0.484102, -0.120601, 0.204408, 0.178344, -0.324401, -0.37313, -0.378014, 0.246637, 0.332265, -0.353111, 0.459299, 0.116516, -0.366959, 0.339219, -0.499194, -0.0738351, 0.457666, 0.0298563, 0.220068, 0.0948994, 0.407688, 0.403166, -0.477476, 0.259953, -0.040639, 0.0735834, -0.0689955, -0.126276, -0.372797, 0.102784, 0.499145}, 
-{0, 0, 0, 0, 0, 0, 1, -0.406548, -0.0472015, 0.435986, 0.389452, -0.20455, 0.0475464, 0.0561968, -0.251458, 0.45522, 0.466573, 0.216623, 0.420822, 0.225599, -0.13744, -0.279521, -0.252463, -0.142626, -0.173081, 0.264353, -0.0600449, -0.3962, 0.224111, 0.0290771, -0.0152757, 0.45284, -0.285113, -0.476945, 0.464492, 0.00590368, 0.493163, -0.384463, 0.378201, -0.10172, 0.0119326, 0.08534, -0.288258, 0.4265, -0.34933, -0.136961, -0.0953322, -0.32639, -0.309284, 0.440288}, 
-{0, 0, 0, 0, 0, 0, 0, 1, -0.243749, 0.147804, -0.223727, 0.172702, 0.305772, -0.485628, -0.298744, 0.0866068, -0.148773, 0.262449, 0.302857, -0.44447, 0.197682, -0.235756, -0.329006, -0.263368, -0.173237, 0.46456, -0.311669, 0.470889, 0.340467, -0.0361063, -0.236003, -0.359672, -0.462934, 0.318087, -0.354197, -0.0383286, 0.0834095, -0.0793594, 0.449992, -0.374079, 0.357154, 0.323571, 0.358512, 0.339343, -0.0606372, 0.491933, -0.125824, -0.392001, 0.368238, -0.0277787}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 1, 0.321942, -0.0336342, 0.484244, -0.269814, -0.267691, -0.17828, -0.268243, 0.259033, 0.109747, -0.309731, -0.366669, 0.486435, -0.394522, 0.406424, -0.313532, -0.303646, 0.236827, -0.0195831, 0.269105, 0.262093, 0.230417, 0.121074, -0.251309, 0.342487, 0.283396, 0.392098, -0.396502, -0.219492, 0.375276, -0.0261682, 0.335855, 0.404843, 0.373349, 0.37033, -0.353981, 0.477698, 0.454513, -0.224287, -0.00555843, -0.129726, -0.235545}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.450007, -0.485359, -0.105622, -0.392058, 0.37934, -0.170947, 0.325218, -0.383213, -0.193537, 0.467443, -0.162515, -0.370993, -0.0681011, 0.298615, -0.495197, -0.490597, 0.374371, -0.374393, 0.176358, -0.416525, -0.0640807, 0.0112317, 0.0126778, 0.407333, 0.177038, -0.0513692, 0.474253, -0.155002, 0.110942, 0.482626, -0.420509, 0.269156, 0.064592, -0.377966, 0.465576, 0.276392, 0.398326, 0.16408, 0.171002, -0.143009}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.487287, 0.138726, 0.0444982, 0.0443997, 0.250806, -0.0695249, -0.375534, -0.147891, 0.474345, -0.448115, -0.41151, -0.384249, -0.208071, 0.424528, -0.322896, 0.325699, 0.342048, -0.0268631, 0.232149, 0.410209, 0.22575, 0.499501, -0.0452405, -0.135706, 0.289086, -0.0770129, 0.172312, -0.168449, 0.135393, -0.0782237, 0.437933, 0.385164, 0.0805287, 0.393131, -0.385007, -0.349623, 0.334558, -0.00744556, 0.489834}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.498669, -0.378728, -0.400112, -0.498225, 0.156091, -0.429447, 0.138058, -0.3077, 0.231999, 0.0208636, -0.41926, 0.067714, -0.0211308, 0.427106, 0.0700811, -0.31604, 0.0276289, -0.370633, 0.217831, 0.248428, 0.281665, -0.0519115, 0.329678, -0.359441, 0.155146, -0.45309, 0.49358, 0.352292, -0.270074, 0.421965, -0.435856, -0.212284, 0.0809322, -0.438511, 0.170346, 0.468461, 0.0138078, 0.430273}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.483884, 0.401714, -0.222275, -0.221175, -0.0530905, 0.210876, 0.303297, 0.321636, -0.298886, 0.00903195, 0.246486, -0.307004, -0.313414, 0.0714729, 0.137156, 0.301679, 0.267622, -0.494616, -0.00898932, 0.440542, -0.434424, 0.211574, 0.0571047, -0.466015, -0.425878, 0.213359, -0.177723, 0.271298, 0.185452, 0.271919, -0.0764722, -0.209377, 0.310405, -0.414365, 0.202714, 0.408026, 0.394814}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.208308, -0.310117, 0.444743, 0.469091, -0.351876, 0.275941, -0.277087, 0.497574, 0.031546, -0.470437, 0.205794, 0.383195, 0.194443, -0.204526, 0.198847, 0.150021, 0.238949, -0.414485, 0.401456, 0.267815, -0.0988932, 0.288736, 0.221099, 0.103914, 0.142621, -0.284307, -0.418776, -0.34845, -0.0879851, -0.221053, -0.194871, 0.395993, -0.442012, 0.0816695, -0.219851, -0.132736}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.3744, -0.371444, 0.401075, -0.48817, -0.338247, 0.307851, -0.476823, -0.451175, 0.341564, 0.0129469, 0.132784, -0.321513, -0.348855, 0.138991, -0.0691676, 0.377533, 0.264391, 0.412024, 0.0272414, -0.486619, 0.0770774, -0.0873933, -0.232728, 0.19571, 0.0200805, 0.0720907, -0.00803746, 0.345069, 0.456333, 0.178451, 0.178691, 0.368482, 0.00279602, 0.419729, 0.338733}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.365912, 0.158724, -0.410033, 0.13029, 0.371597, 0.385236, 0.294604, -0.397558, -0.0176667, 0.204195, 0.0284126, -0.287016, -0.00170742, -0.0189632, -0.0698913, -0.233625, -0.390496, 0.380621, 0.338711, -0.296672, -0.395791, 0.312575, 0.396539, -0.267201, 0.385065, -0.079988, -0.0849446, 0.080143, -0.249779, 0.0773095, 0.31171, 0.112625, 0.177852, 0.357267}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.492936, 0.00813917, -0.376443, 0.0693841, -0.3843, 0.0782057, 0.199562, -0.0723526, 0.185563, 0.114583, 0.260503, -0.495762, -0.25957, 0.0246437, -0.469118, 0.0386655, -0.13015, -0.417984, 0.147163, 0.376344, -0.223784, 0.0112352, 0.0307939, 0.0419242, 0.448681, 0.178227, 0.272016, -0.491049, -0.0482723, 0.418516, 0.160578, 0.279991, -0.424935}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.29741, -0.170394, 0.218562, -0.150926, -0.0989646, -0.0411924, -0.191024, -0.289902, 0.361731, 0.0433798, 0.359592, 0.193775, 0.315496, -0.24161, 0.127366, 0.285524, -0.262074, -0.379607, 0.0566733, -0.367859, 0.485237, 0.0881736, 0.165212, -0.394231, 0.0922336, 0.157126, -0.356473, 0.066541, -0.0647029, 0.226907, -0.283574, -0.214419}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.0620622, 0.315121, 0.115289, -0.129105, 0.474655, 0.398343, 0.308669, 0.409116, -0.221159, 0.325309, 0.019916, 0.107955, 0.406824, -0.0355006, 0.0331047, 0.384828, 0.457873, -0.398027, 0.125551, -0.266688, 0.350814, 0.290687, -0.0183645, 0.00816438, 0.0432712, -0.378558, 0.0602546, -0.402729, 0.444912, 0.147577, -0.0637737}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.431463, -0.490377, 0.202311, -0.468885, 0.404461, 0.126916, 0.495945, -0.369235, -0.30102, -0.445939, 0.0330771, -0.13089, 0.0358122, -0.211783, 0.431048, 0.357812, -0.251485, 0.290367, 0.0615684, 0.104231, -0.459727, 0.0150446, 0.165223, -0.0393211, 0.325593, 0.352813, -0.346587, -0.463006, -0.190812, -0.00137237}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.427591, -0.453887, -0.0819412, 0.22292, -0.112028, 0.388597, 0.275377, 0.390521, -0.342569, -0.458383, -0.390676, 0.0986661, 0.0736035, 0.356411, 0.229704, -0.145365, 0.369089, -0.00289738, -0.354625, -0.149484, -0.346528, 0.244926, -0.108819, 0.475208, -0.0960542, 0.0842137, -0.0926635, 0.298805, 0.410217}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.261188, -0.0716241, -0.41015, -0.342859, -0.359676, -0.00755477, -0.329379, 0.489827, 0.0466252, 0.0761909, -0.250279, -0.483488, 0.394753, -0.0658777, 0.311798, -0.0718978, 0.205255, 0.268743, 0.00186065, -0.266776, -0.0747322, -0.229014, 0.35981, 0.221726, 0.494194, 0.302188, 0.381606, -0.0287991}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.417144, 0.00845891, -0.430407, -0.470063, 0.373847, -0.417228, 0.274954, -0.150475, 0.257741, -0.320329, -0.23454, -0.270258, 0.247004, -0.465837, 0.263629, -0.0381396, -0.120141, -0.115678, 0.0558774, 0.407377, -0.0481891, -0.15188, 0.214995, -0.284789, -0.43711, -0.154001, -0.275419}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.439928, 0.28255, 0.107619, 0.483264, -0.380744, -0.394288, -0.39118, 0.086858, -0.277624, -0.483968, -0.363126, -0.406096, -0.306401, -0.339261, -0.215477, 0.438381, -0.134548, 0.448715, 0.387207, -0.495061, 0.1339, -0.288296, -0.148741, 0.134504, -0.431881, 0.211107}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.308326, 0.175549, 0.449409, -0.418143, 0.459798, -0.378825, -0.229811, -0.463509, 0.00859235, 0.0334963, 0.449626, 0.319514, -0.0996409, 0.171873, 0.332705, -0.145904, -0.310117, 0.389829, -0.38261, -0.383371, 0.0208103, -0.456364, -0.415742, 0.49925, 0.122876}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.386288, 0.150167, -0.446048, 0.399498, 0.340687, 0.34885, -0.332601, -0.32994, -0.0920721, -0.273578, -0.195147, 0.304661, -0.250184, -0.47854, -0.180315, 0.344938, 0.275047, -0.0545408, 0.337189, 0.470654, -0.170467, -0.414849, -0.454512, 0.416094}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.245371, 0.393168, 0.379135, -0.324281, 0.174421, -0.368341, 0.00381498, 0.346158, -0.296472, -0.0722429, -0.466968, -0.0159228, 0.209249, -0.47049, -0.118305, -0.0750544, -0.118197, 0.396568, 0.357734, 0.0348356, 0.454377, -0.36071, 0.458754}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.446855, 0.218748, -0.396231, -0.408087, 0.429423, 0.364734, 0.222905, 0.0080809, 0.370403, -0.48602, -0.267032, -0.19843, -0.32429, 0.00685014, 0.0202666, 0.371556, -0.135029, 0.390194, 0.114504, -0.407677, -0.266225, -0.418069}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.423098, -0.0935819, 0.0484486, -0.14416, 0.0903923, 0.178989, 0.294406, 0.288526, 0.032214, -0.0973917, 0.186939, -0.115563, 0.0305111, 0.078515, -0.0250434, -0.0693544, 0.226989, 0.00802793, -0.244686, -0.0957526, 0.203376}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.391478, -0.0222975, 0.403688, 0.277201, -0.0722725, 0.137531, -0.276635, -0.0188882, -0.284215, -0.249207, 0.121355, -0.115861, 0.0264318, -0.243456, -0.2047, -0.406147, -0.482007, 0.0294766, -0.354967, 0.0227768}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.435874, 0.496434, -0.197375, -0.400467, -0.330342, -0.261005, -0.0427603, 0.323018, -0.27621, -0.108273, 0.0845022, -0.323017, -0.0489245, -0.330467, -0.347261, 0.0572376, -0.219223, 0.297221, -0.334114}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.486626, 0.404713, -0.477857, 0.086308, -0.133541, -0.321805, -0.380896, 0.394228, -0.0803399, 0.487537, -0.283151, 0.429918, 0.49489, 0.00156811, -0.0783027, 0.375309, -0.353828, 0.333314}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.494528, 0.000505648, -0.374533, -0.484144, 0.425859, 0.160632, -0.0678259, -0.453796, 0.269257, 0.112558, 0.409993, 0.282463, -0.0680214, 0.39244, 0.487509, -0.270955, -0.170739}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.371736, 0.472433, -0.498787, 0.105345, 0.413657, -0.481861, 0.0136174, -0.23158, -0.154601, 0.203794, 0.256915, 0.04549, -0.27425, 0.104377, -0.0325464, 0.120376}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.468374, 0.4197, 0.448381, -0.133023, -0.328542, -0.367334, 0.396711, 0.347077, -0.0889141, -0.441881, -0.103937, -0.134741, -0.296802, 0.178283, 0.486032}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.438179, -0.408988, 0.256956, -0.0210645, -0.127653, 0.306739, -0.0127074, -0.479226, 0.237248, -0.146273, 0.493384, -0.392156, 0.386205, 0.140859}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.472394, 0.416662, -0.0744141, 0.470312, 0.415775, 0.280123, 0.0269975, -0.163442, 0.0584176, -0.359572, 0.136424, -0.123553, 0.130225}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.234963, 0.463918, 0.482872, 0.0588538, -0.16535, -0.311888, 0.403967, 0.130383, 0.215229, -0.0746386, 0.346026, -0.441808}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.0254302, -0.404847, -0.0539171, 0.40286, 0.296665, 0.490335, 0.344047, -0.395451, 0.215933, 0.471227, -0.262301}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.385388, 0.453435, 0.356585, -0.394426, 0.395368, 0.241894, -0.0491274, -0.151775, 0.0951899, 0.196336}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.476765, -0.364589, 0.158541, -0.0724698, -0.319819, 0.484328, -0.376204, -0.442835, 0.0695276}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.451903, -0.079951, 0.457803, -0.220424, -0.123529, 0.197099, -0.291604, 0.296917}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.381014, 0.190119, 0.354539, -0.480682, 0.0561141, -0.193193, -0.238729}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.467125, -0.0672749, 0.336962, 0.451713, 0.347155, -0.0944391}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.371928, -0.189838, -0.429483, 0.447512, -0.0371978}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.488108, -0.408249, 0.477063, 0.114118}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.391074, 0.0914219, 0.159453}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0.361859, 0.0914158}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, -0.279576}, 
-{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1}}};
-
-const std::array<float, 50> test_rdiag_compare = {{1309.45, 1280.22, 1087.77, 1121.78, 985.957, 992.468, 944.265, 910.644, 849.876, 949.041, 945.135, 712.22, 627.51, 544.287, 535.576, 462.895, 466.108, 361.608, 370.059, 382.702, 322.907, 303.222, 306.542, 299.656, 262.3, 242.139, 236.443, 227.049, 185.36, 193.295, 193.654, 163.117, 134.461, 109.691, 94.2293, 80.3918, 100.163, 82.1232, 83.3143, 84.1402, 71.2129, 61.3392, 60.0194, 50.873, 48.566, 44.8238, 40.6453, 44.0065, 38.0199, 45.4768}};
