--- fplll/enum-cuda/testdata.h
@@ -1,13 +1,0 @@
-#pragma once
-#include <array>
-
-extern const std::array<std::array<float, 20>, 20> test_mu_small;
-extern const std::array<float, 20> test_solution_small;
-extern const std::array<std::array<float, 50>, 50> test_mu_normal;
-extern const std::array<float, 50> test_solution_normal;
-extern const std::array<std::array<float, 60>, 60> test_mu_big;
-extern const std::array<float, 60> test_solution_big;
-extern const std::array<std::array<float, 50>, 50> test_mu_knapsack_big;
-extern const std::array<std::array<float, 40>, 40> test_mu_knapsack_normal;
-extern const std::array<std::array<float, 50>, 50> test_mu_compare;
-extern const std::array<float, 50> test_rdiag_compare;
