+++ fplll/enum/enumerate.h
@@ -21,36 +21,62 @@
 
 #include <array>
 #include <fplll/enum/enumerate_base.h>
-#include <fplll/enum/enumerate_dyn.h>
 #include <fplll/enum/enumerate_ext.h>
 #include <fplll/enum/evaluator.h>
 #include <fplll/gso_interface.h>
 #include <memory>
-#ifdef FPLLL_WITH_CUDA_ENUM
-#include <fplll/enum-cuda/enumerate_cuda.h>
-#endif
 
 FPLLL_BEGIN_NAMESPACE
 
-template <typename ZT, typename FT> class Enumeration
+template <typename ZT, typename FT> class EnumerationDyn : public EnumerationBase
 {
-private:
-
-  inline void init_external_enumerator() {
-    if (enumext.get() != nullptr) {
-      return;
-    }
-    #ifdef FPLLL_WITH_CUDA_ENUM
-      enumext.reset(new CudaEnumeration<ZT, FT>(_gso, _evaluator));
-    #else
-    #error "expecting cuda to work"
-      if (get_external_enumerator() != nullptr)
-      {
-          enumext.reset(new ExternalEnumeration<ZT, FT>(_gso, _evaluator));
-      }
-    #endif
+public:
+  EnumerationDyn(MatGSOInterface<ZT, FT> &gso, Evaluator<FT> &evaluator,
+                 const vector<int> &max_indices = vector<int>())
+      : _gso(gso), _evaluator(evaluator)
+  {
+    _max_indices = max_indices;
+    std::fill(nodes.begin(), nodes.end(), 0);
   }
 
+  void enumerate(int first, int last, FT &fmaxdist, long fmaxdistexpo,
+                 const vector<FT> &target_coord = vector<FT>(),
+                 const vector<enumxt> &subtree  = vector<enumxt>(),
+                 const vector<enumf> &pruning = vector<enumf>(), bool dual = false,
+                 bool subtree_reset = false);
+
+  inline uint64_t get_nodes(const int level = -1) const
+  {
+    if (level == -1)
+    {
+      return std::accumulate(nodes.cbegin(), nodes.cend(), 0);
+    }
+    return nodes[level];
+  }
+
+  inline array<uint64_t, FPLLL_MAX_ENUM_DIM> get_nodes_array() { return nodes; }
+
+private:
+  MatGSOInterface<ZT, FT> &_gso;
+  Evaluator<FT> &_evaluator;
+  vector<FT> target;
+
+  vector<enumf> pruning_bounds;
+  enumf maxdist;
+  vector<FT> fx;
+
+  void prepare_enumeration(const vector<enumxt> &subtree, bool solvingsvp, bool subtree_reset);
+
+  void do_enumerate();
+
+  void set_bounds();
+  void reset(enumf cur_dist, int cur_depth);
+  virtual void process_solution(enumf newmaxdist);
+  virtual void process_subsolution(int offset, enumf newdist);
+};
+
+template <typename ZT, typename FT> class Enumeration
+{
 public:
   Enumeration(MatGSOInterface<ZT, FT> &gso, Evaluator<FT> &evaluator,
               const vector<int> &max_indices = vector<int>())
@@ -64,10 +90,11 @@ public:
                  const vector<enumf> &pruning = vector<enumf>(), bool dual = false,
                  bool subtree_reset = false)
   {
-    init_external_enumerator();
     // check for external enumerator and use that
-    if (enumext != nullptr && subtree.empty() && target_coord.empty())
+    if (get_external_enumerator() != nullptr && subtree.empty() && target_coord.empty())
     {
+      if (enumext.get() == nullptr)
+        enumext.reset(new ExternalEnumeration<ZT, FT>(_gso, _evaluator));
       if (enumext->enumerate(first, last, fmaxdist, fmaxdistexpo, pruning, dual))
       {
         _nodes = enumext->get_nodes_array();
