--- fplll/mybench.cpp
@@ -1,71 +1,0 @@
-#include "fplll.h"
-#include <config.h>
-#include <array>
-#include "iostream"
-
-using namespace fplll;
-
-int main() {
-  constexpr unsigned int d = 50;
-  const std::array<std::array<float, d>, d> lattice = {{{-1.55418789e+02, -4.34760819e+01, -1.86721311e+01, -3.69582084e+01, 7.77769539e+01, 2.33047756e+01, -5.23553172e+01, 3.15277196e-01, -4.20155120e+01, -7.39743249e+01, 7.48815511e+01, -1.84662358e+00, 2.99835047e+00, 5.79723987e+01, -5.08818789e+01, -7.54670659e+01, -7.31829148e+01, 6.60087500e+01, -9.67707964e+00, 6.35573090e+01, 3.52080983e+01, -7.58981592e+01, -5.48196267e+01, 5.74061866e+01, -3.35866727e+01, 3.48477815e+01, -2.67020482e+00, 5.12614983e+01, -3.88434374e+01, -1.84018935e+01, -2.96424906e+01, -2.52800837e+01, 6.15626982e+01, 6.71347400e+01, 5.50834301e+01, -5.21429877e+01, 1.83118143e+01, 1.23344160e+01, -6.60730923e+01, 2.16190077e+01, -5.40475193e+01, 3.21003659e+01, 3.20746290e+01, 1.96115284e+01, -5.77472007e+01, -4.86363331e+01, 5.07467601e+01, -2.21273118e+01, -3.52402694e+01, 4.58117068e+00},
-  { 0.00000000e+00, 1.53674430e+02, -6.42773891e+01, -7.10710176e+01, 3.09839264e+01, -7.40838909e+01, 6.91708827e+01, 3.23220342e+00, 4.46029322e+01, -3.80487100e+01, 2.10285892e+01, -4.56958518e+01, -6.52590252e+01, -2.38529420e+01, -6.18720028e+01, 2.53652327e+01, -2.53634024e+01, 1.36899927e+01, 2.19768409e+01, -1.94487609e+01, 6.95152095e+01, 6.39289532e+01, -6.12811292e+01, -1.19486628e+01, 5.97938323e+01, -6.46792704e+00, 7.80285678e-01, -1.90490435e+01, -1.61820055e+01, 3.50218171e+01, 1.28405269e+01, 2.47856525e+01, -4.22548833e+01, 3.75843624e+01, 1.81605471e+01, 4.36378857e+01, 6.04207607e+01, 2.49392226e+00, -1.22075724e+00, 9.85791686e+00, 6.65209082e-01, -1.39476806e+01, 2.01561132e+01, -4.83773885e+01, -4.71296496e+01, 4.68814676e+01, 5.78773601e+01, 7.42413114e+01, 7.01412177e+00, -2.64899545e+01},
-  { 0.00000000e+00, 0.00000000e+00, -1.41653693e+02, 5.54334150e+01, 5.71335095e+01, 2.12826732e+01, -5.21874032e+01, -5.01677350e+00, -4.55013966e+01, 5.68237939e+00, 3.82703686e+01, -2.54303963e+01, 5.06424789e+01, 5.92270940e+01, -6.58998427e+01, -5.62165363e+01, -6.77372084e+01, -6.21126130e+01, 5.88553504e+01, -3.57959946e+01, 1.24482007e+01, -5.67087638e+01, -3.75451543e+01, -6.68451827e+01, 3.64462318e+01, 4.15453984e+01, 6.59851267e+00, -4.98380132e+01, 3.06552319e+01, -2.22268007e+01, 3.46064604e+01, 1.16896070e+01, 8.64974326e-01, 9.29214387e-01, -5.25142509e+01, -4.78795047e+01, 1.35075240e+01, -4.86299544e+01, 4.41089278e+01, -1.69649020e+00, 6.02414542e+01, 5.08009223e+01, -4.03059312e+01, 3.85615304e+01, 3.67207802e+01, -4.74203969e+01, -6.99999804e+01, -5.50276764e+01, -3.69551855e+01, 3.49314217e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.43851096e+02, -5.66728426e+01, 7.34690379e-01, 5.67391418e+01, -5.12506694e+01, 3.64094860e+01, 6.26141407e+01, 1.64368593e+00, -1.81770956e+01, -5.23423387e+01, -5.85406944e+01, -4.63273777e+01, 1.16239647e+01, -3.75846463e+01, 4.41678171e+00, -3.23439723e+01, 2.60685169e+01, 2.94700665e+01, -7.78556541e-01, -1.91719687e+01, 7.73798108e+00, 4.70558378e+01, 6.99177086e+00, -3.94921678e+01, -1.98148612e+01, -1.37656478e+00, -3.25130831e+01, -6.30492360e+01, 5.52777151e+01, -5.99661820e+01, 4.54440120e+01, 3.05349240e+01, -3.31759864e+01, 1.12118116e+01, 3.79571112e+01, -5.10401946e+01, 6.44509917e+01, 4.48978987e+01, 7.18518854e+01, -8.67029691e+00, 6.77438051e+01, -2.10402048e+01, -3.29999323e+01, 6.46136879e+01, 6.87656091e+01, -3.65352495e+01, 2.79505981e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.34861754e+02, 6.55299310e+01, -3.49993186e+01, -5.88857616e+01, -3.22919433e+01, -5.99448168e+01, -4.68704023e+01, -3.09001007e+01, 1.26425273e+01, -5.20430863e+01, 2.09246853e+01, 3.02139393e+01, -4.46445156e+01, -4.51565675e+01, -3.79876965e+01, 5.60734747e+01, 2.22768104e+01, -6.58763887e+01, -3.27074942e+01, 4.47274960e+01, 2.63152182e+00, -9.62953952e+00, 3.31793513e+01, 4.64571514e+01, -2.32612389e+01, 6.95218525e+00, 4.29381577e+01, -1.85384512e+01, -2.80301118e+01, 1.06549587e+00, -6.02981771e+01, 5.15728395e+01, 5.55949802e+01, 2.66203527e+01, 1.66904477e+01, 5.42296220e+00, -5.13097766e+01, -3.72018062e+01, -6.64922870e+01, -1.14982557e+01, 1.75537012e+00, -2.44124633e+01, 1.56308984e+01, 3.08480757e+01, -5.62337915e+00, 5.72894613e-01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.35306322e+02, -4.38850026e+01, -1.75135175e+01, -9.85725331e+00, 3.20044177e+01, -4.06448648e+01, -1.56964472e+01, -4.14048225e+01, 2.36480138e+01, -1.11344393e+01, 1.46473493e+01, -5.12889129e+01, 3.56352385e+01, -1.62287554e+01, -6.55020205e+01, -1.63180834e+01, 2.76576615e+01, 2.41310318e+01, -4.38935315e+01, -5.04869133e+01, -5.11476455e+01, 3.33715010e+01, 4.49575817e+01, -4.77782176e+01, 6.21461061e+01, 1.57653901e+01, -4.96518590e+01, 4.58984310e+01, -6.75441431e+01, -9.99035122e+00, 6.19251058e+01, 4.03974969e+00, 2.97766390e+01, 1.28404873e+01, 5.51627935e+01, 5.45509431e+01, -6.46054572e+01, 3.51733320e+01, -5.49871355e+00, 9.95629844e+00, -9.33552137e+00, -1.70859560e+01, -5.04418442e+01, 1.39073094e+01, 6.75374735e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.31979577e+02, -5.36560562e+01, -6.22963027e+00, 5.75412955e+01, 5.13997630e+01, -2.69963950e+01, 6.27515198e+00, 7.41682901e+00, -3.31873093e+01, 6.00797479e+01, 6.15780431e+01, 2.85897509e+01, 5.55398856e+01, 2.97744765e+01, -1.81392475e+01, -3.68911243e+01, -3.33200175e+01, -1.88236708e+01, -2.28431068e+01, 3.48891585e+01, -7.92469907e+00, -5.22903242e+01, 2.95781404e+01, 3.83758028e+00, -2.01607931e+00, 5.97656313e+01, -3.76290752e+01, -6.29470659e+01, 6.13034236e+01, 7.79165349e-01, 6.50874065e+01, -5.07413159e+01, 4.99147644e+01, -1.34249624e+01, 1.57486310e+00, 1.12631335e+01, -3.80442075e+01, 5.62892626e+01, -4.61044005e+01, -1.80760668e+01, -1.25819027e+01, -4.30767518e+01, -4.08191721e+01, 5.81090099e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.29608741e+02, 3.15920601e+01, -1.91567338e+01, 2.89970308e+01, -2.23836568e+01, -3.96307202e+01, 6.29415863e+01, 3.87198130e+01, -1.12249957e+01, 1.92822532e+01, -3.40156919e+01, -3.92529488e+01, 5.76071316e+01, -2.56212832e+01, 3.05560380e+01, 4.26420507e+01, 3.41348134e+01, 2.24530098e+01, -6.02110041e+01, 4.03949711e+01, -6.10313423e+01, -4.41275517e+01, 4.67969189e+00, 3.05880733e+01, 4.66166049e+01, 6.00002727e+01, -4.12269033e+01, 4.59070293e+01, 4.96771950e+00, -1.08105986e+01, 1.02856736e+01, -5.83229079e+01, 4.84838746e+01, -4.62902384e+01, -4.19376422e+01, -4.64662516e+01, -4.39818082e+01, 7.85911457e+00, -6.37588379e+01, 1.63078747e+01, 5.08067197e+01, -4.77268793e+01, 3.60036799e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.25209608e+02, -4.03102064e+01, 4.21132288e+00, -6.06319480e+01, 3.37833165e+01, 3.35175154e+01, 2.23223596e+01, 3.35865589e+01, -3.24334609e+01, -1.37414043e+01, 3.87812411e+01, 4.59105100e+01, -6.09063723e+01, 4.93979200e+01, -5.08881354e+01, 3.92571700e+01, 3.80193844e+01, -2.96530531e+01, 2.45199401e+00, -3.36945184e+01, -3.28165106e+01, -2.88504102e+01, -1.51595660e+01, 3.14662639e+01, -4.28826008e+01, -3.54839241e+01, -4.90943917e+01, 4.96458997e+01, 2.74825363e+01, -4.69881222e+01, 3.27650668e+00, -4.20523250e+01, -5.06902099e+01, -4.67468453e+01, -4.63688511e+01, 4.43218214e+01, -5.98123372e+01, -5.69093890e+01, 2.80828719e+01, 6.95969177e-01, 1.62429320e+01, 2.94924660e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.32312952e+02, 5.95417566e+01, 6.42192939e+01, 1.39751387e+01, 5.18744011e+01, -5.01915856e+01, 2.26185131e+01, -4.30306097e+01, 5.07039796e+01, 2.56074890e+01, -6.18487469e+01, 2.15028202e+01, 4.90871795e+01, 9.01065878e+00, -3.95106322e+01, 6.55209292e+01, 6.49123167e+01, -4.95341031e+01, 4.95369853e+01, -2.33344882e+01, 5.51117098e+01, 8.47870909e+00, -1.48609762e+00, -1.67744045e+00, -5.38954510e+01, -2.34243906e+01, 6.79681486e+00, -6.27498782e+01, 2.05087193e+01, -1.46790404e+01, -6.38576870e+01, 5.56387341e+01, -3.56128668e+01, -8.54635429e+00, 5.00098424e+01, -6.16017883e+01, -3.65702659e+01, -5.27036877e+01, -2.17099190e+01, -2.26257886e+01, 1.89219597e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.32040410e+02, 6.43416336e+01, -1.83173787e+01, -5.87556413e+00, -5.86254799e+00, -3.31164773e+01, 9.18010138e+00, 4.95856529e+01, 1.95275615e+01, -6.26327605e+01, 5.91692265e+01, 5.43359155e+01, 5.07363835e+01, 2.74737780e+01, -5.60548830e+01, 4.26353795e+01, -4.30053987e+01, -4.51642048e+01, 3.54701800e+00, -3.06530885e+01, -5.41642277e+01, -2.98080709e+01, -6.59542929e+01, 5.97357163e+00, 1.79187344e+01, -3.81709802e+01, 1.01688168e+01, -2.27521374e+01, 2.22420928e+01, -1.78773465e+01, 1.03286898e+01, -5.78248934e+01, -5.08571756e+01, -1.06330360e+01, -5.19092226e+01, 5.08364692e+01, 4.61643053e+01, -4.41751442e+01, 9.83114206e-01, -6.46778192e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.14621737e+02, 5.71583182e+01, 4.34104110e+01, 4.58615505e+01, 5.71073634e+01, -1.78913724e+01, 4.92239088e+01, -1.58244225e+01, 3.52691239e+01, -2.65920813e+01, -2.39142630e+00, 4.80563207e+01, -7.76149463e+00, 2.42204438e+00, -4.89555734e+01, -8.03281234e+00, 3.62251040e+01, -3.16686680e+00, 4.24826534e+01, -2.49681136e+01, -2.84752643e+01, -3.22849443e+01, 5.95018337e+00, -3.77882296e+01, 4.11997886e+01, -1.77831203e+01, 5.19339525e+01, -5.65749642e+01, -4.03803614e+01, 3.09563648e+01, -4.83663219e+01, 4.99585727e+01, 2.43323464e+01, -9.27659215e+00, 5.02629397e+01, -1.95253086e+01, -5.36958471e+01, -1.58267019e+00, -4.93186617e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -1.07589532e+02, 5.20608659e+01, -4.32202617e+01, 2.39144702e+01, 2.37961541e+01, 5.71198276e+00, -2.26880284e+01, -3.26315881e+01, -3.46047167e+01, 3.21569677e+01, -9.71743125e-01, -2.65192596e+01, 3.30303920e+01, 3.37200463e+01, -7.68973953e+00, -1.47565606e+01, -3.24575364e+01, -2.87932790e+01, 5.32154896e+01, 9.67156203e-01, -4.73977210e+01, 4.67395007e+01, -2.27631874e+01, -6.14387210e+00, 5.01383546e+01, 4.58200246e+01, -2.29552104e+01, 1.91211770e+01, -2.91888367e+01, -1.99526548e+01, -2.92555987e+01, 8.22760670e+00, 2.25267862e+01, -3.33962758e+01, 4.45813781e+01, -2.18099011e+01, -4.38993246e+01, -4.24778427e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 1.00201374e+02, -2.08727331e+01, -3.10741935e+01, 4.45638102e+01, 4.70035527e+01, -3.52584557e+01, 2.76496624e+01, -2.77645142e+01, 4.98576405e+01, 3.16095529e+00, -4.71383868e+01, 2.06208533e+01, 3.83967044e+01, 1.94834426e+01, -2.04937937e+01, 1.99247015e+01, 1.50322634e+01, 2.39429737e+01, -4.15319707e+01, 4.02264770e+01, 2.68353846e+01, -9.90923259e+00, 2.89317904e+01, 2.21544392e+01, 1.04123301e+01, 1.42908607e+01, -2.84879653e+01, -4.19619221e+01, -3.49151544e+01, -8.81622836e+00, -2.21497863e+01, -1.95262923e+01, 3.96790399e+01, -4.42901616e+01, 8.18339970e+00, -2.20293397e+01, -1.33003653e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -9.93963711e+01, 3.72140222e+01, 3.69201454e+01, -3.98654359e+01, 4.85222805e+01, 3.36205147e+01, -3.05993083e+01, 4.73944586e+01, 4.48451444e+01, -3.39502469e+01, -1.28687184e+00, -1.31982236e+01, 3.19571811e+01, 3.46749620e+01, -1.38152003e+01, 6.87500378e+00, -3.75254393e+01, -2.62795031e+01, -4.09536492e+01, -2.70769598e+00, 4.83682110e+01, -7.66121266e+00, 8.68657598e+00, 2.31323643e+01, -1.94528600e+01, -1.99592960e+00, -7.16555785e+00, 7.98894457e-01, -3.42986211e+01, -4.53578787e+01, -1.77374000e+01, -1.77611935e+01, -3.66257643e+01, -2.77914599e-01, -4.17195107e+01, -3.36688575e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 9.24062275e+01, 3.38125127e+01, 1.46670782e+01, -3.78895617e+01, 1.20396419e+01, 3.43379214e+01, 3.55982262e+01, 2.72232301e+01, -3.67368676e+01, -1.63251047e+00, 1.88688586e+01, 2.62550504e+00, -2.65220449e+01, -1.57775863e-01, -1.75231547e+00, -6.45838917e+00, -2.15884416e+01, -3.60842473e+01, 3.51717559e+01, 3.12989981e+01, -2.74143645e+01, -3.65735285e+01, 2.88838411e+01, 3.66427136e+01, -2.46910522e+01, 3.55823606e+01, -7.39139150e+00, -7.84941149e+00, 7.40571390e+00, -2.30811252e+01, 7.14388357e+00, 2.88039031e+01, 1.04072886e+01, 1.64346170e+01, 3.30137271e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 9.27262973e+01, -4.57081377e+01, 7.54714985e-01, -3.49062063e+01, 6.43372916e+00, -3.56347096e+01, 7.25172744e+00, 1.85046848e+01, -6.70898787e+00, 1.72065802e+01, 1.06248116e+01, 2.41554337e+01, -4.59701677e+01, -2.40689732e+01, 2.28511902e+00, -4.34996164e+01, 3.58530819e+00, -1.20683726e+01, -3.87580860e+01, 1.36458617e+01, 3.48969496e+01, -2.07506701e+01, 1.04179491e+00, 2.85540117e+00, 3.88747165e+00, 4.16045560e+01, 1.65263497e+01, 2.52230744e+01, -4.55331333e+01, -4.47610871e+00, 3.88074015e+01, 1.48898320e+01, 2.59625160e+01, -3.94026393e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -8.16731092e+01, -2.42904062e+01, 1.39166211e+01, -1.78506042e+01, 1.23265781e+01, 8.08275006e+00, 3.36430965e+00, 1.56014921e+01, 2.36771823e+01, -2.95436879e+01, -3.54296000e+00, -2.93689661e+01, -1.58262304e+01, -2.57675224e+01, 1.97330386e+01, -1.04023409e+01, -2.33196476e+01, 2.14044342e+01, 3.10036722e+01, -4.62868694e+00, 3.00442088e+01, -3.96308387e+01, -7.20141350e+00, -1.34933394e+01, 3.21980954e+01, -7.53300344e+00, -1.28329820e+01, 2.91142211e+01, -5.43460634e+00, 5.28449068e+00, -1.85322301e+01, 2.31603678e+01, 1.75122670e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -8.26219209e+01, 5.12769761e+00, -2.60359319e+01, -9.52541952e+00, 1.06669092e+01, -3.92168641e+01, -3.29118535e+01, -2.55028082e+01, -3.38019375e+01, 1.82725412e+01, -2.68776554e+01, -1.64550163e+00, -8.91948299e+00, -3.36125792e+01, 2.93312598e+00, -2.73517597e+00, -3.17952447e+01, -3.78303121e+01, 3.28857427e+01, -1.03732636e+01, 2.20342879e+01, -2.89849481e+01, -2.40171264e+01, 1.51730697e+00, -6.74556948e-01, -3.57514853e+00, 3.12771600e+01, -4.97835033e+00, 3.32742408e+01, -3.67595230e+01, -1.21931035e+01, 5.26910525e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -8.40214355e+01, -3.62521298e+01, 4.12021620e+01, -1.69984913e+01, 3.93964074e+01, -3.39834262e+01, -1.06636837e+01, -4.16699967e+01, 3.10236420e+01, 2.52921022e+01, 3.74684020e+01, -2.77918680e+00, 1.09975456e+01, -3.00899308e+00, 1.77943101e+01, -3.62172970e+01, -3.00638565e+01, 2.11301672e+01, -2.43970264e+01, -5.17306557e+00, -8.75765517e+00, 3.86269068e+01, -1.26406945e+00, -1.38822714e+01, 3.30381135e+00, -2.73568070e+01, -2.96438367e+01, 2.91207075e+01, 3.89024238e+01, 1.60322915e+01, 1.15308696e-01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -7.71788958e+01, 3.30009699e+01, 3.50305321e+01, 6.32412805e+00, -1.72047261e+01, 8.64616823e+00, -2.99915070e+01, -2.12533048e+01, -3.01399535e+01, 2.64390758e+01, 3.53774938e+01, 3.01519789e+01, -7.61494281e+00, -5.68063359e+00, -2.75074139e+01, -1.77282883e+01, 1.12191428e+01, -2.84858458e+01, 2.23616434e-01, 2.73695900e+01, 1.15370267e+01, 2.67446673e+01, -1.89030951e+01, 8.39852149e+00, -3.66760514e+01, 7.41335778e+00, -6.49951852e+00, 7.15166811e+00, -2.30614233e+01, -3.16601225e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 7.47894680e+01, -1.95340898e+01, -5.35672794e+00, -3.06748646e+01, -2.56422263e+01, -2.68999880e+01, -5.65016905e-01, -2.46340628e+01, 3.66339054e+01, 3.48707112e+00, 5.69827366e+00, -1.87182566e+01, -3.61597792e+01, 2.95233308e+01, -4.92695854e+00, 2.33191809e+01, -5.37719488e+00, 1.53509154e+01, 2.00991560e+01, 1.39157296e-01, -1.99520409e+01, -5.58918463e+00, -1.71277988e+01, 2.69100244e+01, 1.65827970e+01, 3.69604750e+01, 2.26004636e+01, 2.85400843e+01, -2.15387194e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -7.51976921e+01, -3.13682364e+01, -6.36090455e-01, 3.23656383e+01, 3.53476416e+01, -2.81124000e+01, 3.13745758e+01, -2.06759153e+01, 1.13153858e+01, -1.93815564e+01, 2.40880008e+01, 1.76368427e+01, 2.03227763e+01, -1.85741109e+01, 3.50298452e+01, -1.98243113e+01, 2.86801373e+00, 9.03431925e+00, 8.69874540e+00, -4.20185261e+00, -3.06338290e+01, 3.62371197e+00, 1.14209933e+01, -1.61671639e+01, 2.14154405e+01, 3.28696589e+01, 1.15805139e+01, 2.07108404e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 7.43483731e+01, -3.27079619e+01, 2.10071511e+01, 8.00126339e+00, 3.59298996e+01, -2.83076734e+01, -2.93146514e+01, -2.90836132e+01, 6.45775167e+00, -2.06409279e+01, -3.59821957e+01, -2.69978161e+01, -3.01926122e+01, -2.27804311e+01, -2.52234716e+01, -1.60203310e+01, 3.25929328e+01, -1.00034429e+01, 3.33611961e+01, 2.87881729e+01, -3.68069897e+01, 9.95521359e+00, -2.14343232e+01, -1.10586826e+01, 1.00001547e+01, -3.21096299e+01, 1.56954461e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.95598997e+01, 2.14470959e+01, -1.22112037e+01, -3.12608666e+01, 2.90859503e+01, -3.19834736e+01, 2.63510004e+01, 1.59856272e+01, 3.22416118e+01, -5.97683247e-01, -2.33000210e+00, -3.12759451e+01, -2.22253291e+01, 6.93101257e+00, -1.19554985e+01, -2.31429255e+01, 1.01490400e+01, 2.15717389e+01, -2.71164742e+01, 2.66143151e+01, 2.66672686e+01, -1.44755918e+00, 3.17446153e+01, 2.89189441e+01, -3.47277624e+01, -8.54724857e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.68332092e+01, 2.58168766e+01, -1.00361517e+01, 2.98107937e+01, -2.66997626e+01, -2.27692036e+01, -2.33147326e+01, 2.22287764e+01, 2.20509292e+01, 6.15347418e+00, 1.82841085e+01, 1.30422777e+01, -2.03614866e+01, 1.67206197e+01, 3.19823686e+01, 1.20510602e+01, -2.30533267e+01, -1.83822973e+01, 3.64513612e+00, -2.25354363e+01, -3.14552951e+01, 1.13928789e+01, 2.77257014e+01, 3.03765045e+01, -2.78089042e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.60424938e+01, 1.62049018e+01, -2.59657845e+01, -2.50390446e+01, 2.14163557e+01, -1.15191726e+01, 2.43261406e+01, -2.51950726e-01, -2.28611232e+01, 1.95797696e+01, 4.77109895e+00, 3.08397424e+01, 1.05158035e+00, -1.38193487e+01, 3.10723471e+01, 7.81314690e+00, 4.95678227e+00, 7.80605133e+00, -2.61903422e+01, -2.36256375e+01, -2.30063300e+00, -3.00082180e+01, 2.38221884e+01, -3.02972546e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -6.47172349e+01, -2.89192293e+01, -1.41567814e+01, 2.56429931e+01, 2.64102857e+01, -2.77910382e+01, -2.36045934e+01, -1.44257935e+01, -5.22973532e-01, -2.39714388e+01, 3.14538848e+01, 1.72815913e+01, 1.28418388e+01, 2.09871486e+01, -4.43321806e-01, -1.31159515e+00, -2.40460862e+01, 8.73870306e+00, -2.52522653e+01, -7.41038373e+00, 2.63837047e+01, 1.72293656e+01, 2.70562510e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.84746613e+01, -2.47405055e+01, -5.47216848e+00, 2.83301761e+00, -8.42969519e+00, 5.28566216e+00, 1.04663334e+01, 1.72153057e+01, 1.68714824e+01, 1.88370092e+00, -5.69494702e+00, 1.09311757e+01, -6.75753357e+00, 1.78412854e+00, 4.59113789e+00, -1.46440300e+00, -4.05547245e+00, 1.32730796e+01, 4.69430327e-01, -1.43079381e+01, -5.59910100e+00, 1.18923512e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.97131168e+01, 2.33763537e+01, -1.33145194e+00, 2.41054494e+01, 1.65525243e+01, -4.31561856e+00, 8.21238839e+00, -1.65187535e+01, -1.12787205e+00, -1.69713869e+01, -1.48809517e+01, 7.24650440e+00, -6.91839594e+00, 1.57832731e+00, -1.45375433e+01, -1.22232694e+01, -2.42522941e+01, -2.87821537e+01, 1.76014003e+00, -2.11961991e+01, 1.36007523e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -5.97685926e+01, -2.60515567e+01, -2.96711898e+01, 1.17968506e+01, 2.39353511e+01, 1.97440563e+01, 1.55998902e+01, 2.55572114e+00, -1.93063184e+01, 1.65086760e+01, 6.47133121e+00, -5.05057999e+00, 1.93062575e+01, 2.92415119e+00, 1.97515360e+01, 2.07552921e+01, -3.42101219e+00, 1.31026782e+01, -1.77644659e+01, 1.99694988e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 5.48541057e+01, -2.66934300e+01, 2.22001752e+01, -2.62124431e+01, 4.73434761e+00, -7.32526779e+00, -1.76523075e+01, -2.08937330e+01, 2.16250025e+01, -4.40697368e+00, 2.67434082e+01, -1.55319747e+01, 2.35827561e+01, 2.71467698e+01, 8.60171818e-02, -4.29522644e+00, 2.05872503e+01, -1.94089029e+01, 1.82836179e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.98033603e+01, -2.46291452e+01, -2.51829880e-02, 1.86530176e+01, 2.41120076e+01, -2.12092029e+01, -8.00000272e+00, 3.37795979e+00, 2.26005899e+01, -1.34099211e+01, -5.60574359e+00, -2.04190241e+01, -1.40676173e+01, 3.38769259e+00, -1.95448255e+01, -2.42795907e+01, 1.34944729e+01, 8.50336639e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.49827111e+01, 1.67216897e+01, -2.12513246e+01, 2.24368129e+01, -4.73868443e+00, -1.86074331e+01, 2.16754130e+01, -6.12549324e-01, 1.04171051e+01, 6.95436608e+00, -9.16721941e+00, -1.15567549e+01, -2.04626213e+00, 1.23365240e+01, -4.69516047e+00, 1.46402455e+00, -5.41485846e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 4.16920001e+01, -1.95274645e+01, 1.74981292e+01, 1.86938890e+01, -5.54600820e+00, -1.36975705e+01, -1.53149083e+01, 1.65396631e+01, 1.44703546e+01, -3.70700517e+00, -1.84229033e+01, -4.33333945e+00, -5.61760568e+00, -1.23742864e+01, 7.43297751e+00, 2.02636465e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.85093194e+01, -1.68739583e+01, 1.57498315e+01, -9.89521943e+00, 8.11179022e-01, 4.91582051e+00, -1.18123065e+01, 4.89352152e-01, 1.84546751e+01, -9.13624691e+00, 5.63286969e+00, -1.89998721e+01, 1.51016538e+01, -1.48724993e+01, -5.42440273e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -4.29846494e+01, -2.03056710e+01, -1.79100607e+01, 3.19866338e+00, -2.02161804e+01, -1.78719397e+01, -1.20409676e+01, -1.16047746e+00, 7.02550238e+00, -2.51106032e+00, 1.54560774e+01, -5.86414956e+00, 5.31087452e+00, -5.59769097e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.89217814e+01, -9.14519734e+00, -1.80565099e+01, -1.87942505e+01, -2.29069300e+00, 6.43570831e+00, 1.21392393e+01, -1.57231093e+01, -5.07475659e+00, -8.37708124e+00, 2.90506746e+00, -1.34679415e+01, 1.71959484e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.92030247e+01, 9.96939826e-01, 1.58712225e+01, 2.11371255e+00, -1.57933187e+01, -1.16301598e+01, -1.92226251e+01, -1.34876777e+01, 1.55028575e+01, -8.46522126e+00, -1.84735066e+01, 1.02830046e+01},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.93968739e+01, -1.51830899e+01, 1.78639381e+01, 1.40483422e+01, -1.55391510e+01, 1.55762783e+01, 9.52986876e+00, -1.93546550e+00, -5.97945107e+00, 3.75018271e+00, 7.73503737e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.62442471e+01, 1.72799801e+01, 1.32142559e+01, -5.74620464e+00, 2.62661386e+00, 1.15916162e+01, -1.75541101e+01, 1.36352469e+01, 1.60502291e+01, -2.51997641e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.36379150e+01, -1.52010726e+01, 2.68938612e+00, -1.53995237e+01, 7.41461917e+00, 4.15527213e+00, -6.63000271e+00, 9.80895488e+00, -9.98768306e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 3.32740560e+01, -1.26778877e+01, 6.32603124e+00, 1.17969463e+01, -1.59942418e+01, 1.86714530e+00, -6.42831084e+00, -7.94348164e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -3.06339887e+01, 1.43098970e+01, 2.06089920e+00, -1.03224755e+01, -1.38377813e+01, -1.06347458e+01, 2.89304780e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.99313298e+01, 1.11322977e+01, 5.68209605e+00, 1.28550011e+01, -1.33946331e+01, 1.11337943e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.87550570e+01, -1.40355694e+01, -1.17392345e+01, 1.37179852e+01, 3.28147346e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.73819930e+01, -1.07083939e+01, -2.50331328e+00, -4.36615067e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 2.84916995e+01, 1.03099884e+01, 2.60459029e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.64828936e+01, 7.40398109e+00},
-  { 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, 0.00000000e+00, -2.89637400e+01}}};
-  ZZ_mat<mpz_t> b(d, d);
-  for (unsigned int i = 0; i < d; ++i) {
-    for (unsigned int j = 0; j < d; ++j) {
-      b(i, j) = lattice[j][i] * 1e9;
-    }
-  }
-  set_threads(-1);
-  std::vector<Z_NR<mpz_t>> sol_coord;
-  shortest_vector(b, sol_coord, SVPM_FAST, SVP_DEFAULT);
-  std::cout << sol_coord << std::endl;
-  return 0;
-}
