+++ tools/plot_gso_dump.py
