+++ tools/reformat_magma.pl
